% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screening_facility.R
\name{allocate_screening_facility}
\alias{allocate_screening_facility}
\title{Allocate the screening facility}
\usage{
allocate_screening_facility(facility_data, research_facilities)
}
\arguments{
\item{facility_data}{dataframe containing the processed facility data}

\item{research_facilities}{dataframe containing the list of research facilities and their characteristics}
}
\description{
Several information can be triangulated to:
the device ID,
the GPS coordinates,
the selection of the facility by the research assistant
and the child ID
}
