% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anonymise_dataframe}
\alias{anonymise_dataframe}
\title{Anonymise a dataframe by hashing selected columns}
\usage{
anonymise_dataframe(df, cols_to_anon, algo = "sha256")
}
\arguments{
\item{df}{Input dataframe}

\item{cols_to_anon}{Columns to use for hash}

\item{algo}{hash functions algorithms to be used, currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, murmur32, spookyhash and blake3.}
}
\value{
This function returns a vector
}
\description{
Anonymise a dataframe by hashing selected columns
}
