% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{calculate_antibio_has_been_recorded}
\alias{calculate_antibio_has_been_recorded}
\title{Create derived variable rx_has_been_recorded (TIMCI-specific function)}
\usage{
calculate_antibio_has_been_recorded(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a dataframe with a new column.
}
\description{
Create derived variable rx_has_been_recorded (TIMCI-specific function)
}
