% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heatmap.R
\name{calendar_heatmap}
\alias{calendar_heatmap}
\title{Calendar Heatmap}
\usage{
calendar_heatmap(dates, values, title = "", subtitle = "", legendtitle = "")
}
\arguments{
\item{dates}{A vector containing the dates in `Date` format.}

\item{values}{A vector containing the corresponding values as numeric.}

\item{title}{Main plot title (optional).}

\item{subtitle}{Main plot subtitle (optional).}

\item{legendtitle}{Legend title (optional).}
}
\value{
ggplot object
}
\description{
Creates a colour coded calendar visualising time series data
}
