% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fu_data.R
\name{clean_followup_for_rate_estimation}
\alias{clean_followup_for_rate_estimation}
\title{Clean follow-up data for estimating rough follow-up rate (TIMCI-specific function)}
\usage{
clean_followup_for_rate_estimation(day0_df, fu_df)
}
\arguments{
\item{day0_df}{dataframe containing the non de-identified (raw) ODK data collected at Day 0}

\item{fu_df}{dataframe containing follow-up data that correspond to successful (either Day 7 Day 28) calls}
}
\value{
out
}
\description{
Clean follow-up data for estimating rough follow-up rate (TIMCI-specific function)
}
