% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_odk_data.R
\name{combine_columns}
\alias{combine_columns}
\title{Combine 2 columns - function in progress (to be tested)}
\usage{
combine_columns(df, cols1, cols2)
}
\arguments{
\item{df}{Dataframe containing the non de-identified (raw) ODK data collected at the facility level}

\item{cols1}{Vector of reference column names}

\item{cols2}{Vector of column names to be merged with the reference columns}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Combine 2 columns - function in progress (to be tested)
}
