% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_dataframes}
\alias{combine_dataframes}
\title{Combine two dataframes in one}
\usage{
combine_dataframes(df1, df2, verbose = FALSE)
}
\arguments{
\item{df1}{Input dataframe 1}

\item{df2}{Input dataframe 2}

\item{verbose}{If set to true, writes text in stderr()}
}
\value{
This function returns a dataframe.
}
\description{
Combine two dataframes in one
}
