% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{concatenate_names}
\alias{concatenate_names}
\title{Concatenate names for fuzzy-matching scoring (TIMCI-specific function)}
\usage{
concatenate_names(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a dataframe with an additional column `name` that contains the concatenated name.
}
\description{
Concatenate names for fuzzy-matching scoring (TIMCI-specific function)
}
