% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{convert_yi_age2ctg}
\alias{convert_yi_age2ctg}
\title{Convert age in days to age categories (TIMCI-specific function)}
\usage{
convert_yi_age2ctg(yi_ctg, yr_ctg, val, incl, excl)
}
\arguments{
\item{yi_ctg}{character containing the category of young infant}

\item{yr_ctg}{scalar value}

\item{val}{scalar value containing the age of the child in days}

\item{incl}{boolean value}

\item{excl}{boolean}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Convert age in days to age categories (TIMCI-specific function)
}
