% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rctls_data.R
\name{correct_day0_all}
\alias{correct_day0_all}
\title{Correct Day 0 data for all errors (TIMCI-specific function)}
\usage{
correct_day0_all(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns an edited dataframe with corrections
}
\description{
Correct Day 0 data for all errors (TIMCI-specific function)
}
