% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_day0_drug_data}
\alias{correct_day0_drug_data}
\title{Edit drug data in Day 0 data entries (TIMCI-specific function - Kenya and Senegal only)}
\usage{
correct_day0_drug_data(day0_df, drug_df)
}
\arguments{
\item{day0_df}{dataframe that contains Day 0 data and needs to be corrected}

\item{drug_df}{dataframe that contains corrected (structured) drug data to edit in Day 0 data}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Edit drug data in Day 0 data entries (TIMCI-specific function - Kenya and Senegal only)
}
