% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rctls_data.R
\name{correct_day0_facilities}
\alias{correct_day0_facilities}
\title{Correct Day 0 facilities (TIMCI-specific function)}
\usage{
correct_day0_facilities(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Correct Day 0 facilities (TIMCI-specific function)
}
