% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_day0_inconsistent_facilities}
\alias{correct_day0_inconsistent_facilities}
\title{Edit non-valid facilities in Day 0 data entries (TIMCI-specific function)}
\usage{
correct_day0_inconsistent_facilities(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Edit non-valid facilities in Day 0 data entries (TIMCI-specific function)
}
