% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_day28_duplicates}
\alias{correct_day28_duplicates}
\title{Correct Day 28 duplicates (TIMCI-specific function)}
\usage{
correct_day28_duplicates(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Correct Day 28 duplicates (TIMCI-specific function)
}
