% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_day7_all}
\alias{correct_day7_all}
\title{Edit Day 7 follow-up data for all errors that were detected by quality checks (TIMCI-specific function)}
\usage{
correct_day7_all(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns an edited dataframe with corrections
}
\description{
Edit Day 7 follow-up data for all errors that were detected by quality checks (TIMCI-specific function)
}
