% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rctls_data.R
\name{correct_duplicates}
\alias{correct_duplicates}
\title{Correct duplicates (TIMCI-specific function)}
\usage{
correct_duplicates(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Correct duplicates (TIMCI-specific function)
}
