% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_spa_sco_all}
\alias{correct_spa_sco_all}
\title{Edit SPA sick child observation data for all errors that were detected by quality checks (TIMCI-specific function)}
\usage{
correct_spa_sco_all(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns an edited dataframe with corrections
}
\description{
Edit SPA sick child observation data for all errors that were detected by quality checks (TIMCI-specific function)
}
