% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_spa_sco_fids}
\alias{correct_spa_sco_fids}
\title{Edit non-valid facilities in SPA sick child observation data entries (TIMCI-specific function)}
\usage{
correct_spa_sco_fids(df, csv_prefix = "spa_sco_facility_correction")
}
\arguments{
\item{df}{dataframe}

\item{csv_prefix}{A string value indicating the prefix of the CSV file from which to read the corrections (default is "spa_sco_facility_correction").}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Edit non-valid facilities in SPA sick child observation data entries (TIMCI-specific function)
}
