% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_spa_sco_hcp_ids}
\alias{correct_spa_sco_hcp_ids}
\title{Edit incorrect healthcare provider (HCP) IDs in SPA sick child observation entries (TIMCI-specific function)
This function can be used to correct documented HCP ID duplicates, incorrect facility codes or typos}
\usage{
correct_spa_sco_hcp_ids(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Edit incorrect healthcare provider (HCP) IDs in SPA sick child observation entries (TIMCI-specific function)
This function can be used to correct documented HCP ID duplicates, incorrect facility codes or typos
}
