% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{correct_spo2_values}
\alias{correct_spo2_values}
\title{Correct hospital follow-up IDs (TIMCI-specific function)}
\usage{
correct_spo2_values(df, csv_prefix = "day0_spo2_meas1", meas = 1)
}
\arguments{
\item{df}{dataframe}

\item{csv_prefix}{A string value indicating the prefix of the CSV file from which to read the corrections (default is "day7_non_valid_pid_correction").}

\item{meas}{integer with value 1 or 2 to indicate the measurement variable to replace}
}
\value{
This function returns a list that contains a dataframe with corrections and the list of edits
}
\description{
Correct hospital follow-up IDs (TIMCI-specific function)
}
