% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{count_baseline_vs_repeat}
\alias{count_baseline_vs_repeat}
\title{Count the occurrence of baseline versus repeat visits (TIMCI-specific function)}
\usage{
count_baseline_vs_repeat(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns de-identified data.
}
\description{
Count the occurrence of baseline versus repeat visits (TIMCI-specific function)
}
