% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{count_noneligibility}
\alias{count_noneligibility}
\title{Count the occurrence of non-eligibility reasons (TIMCI-specific function)}
\usage{
count_noneligibility(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns de-identified data.
}
\description{
Count the occurrence of non-eligibility reasons (TIMCI-specific function)
}
