% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{count_screening}
\alias{count_screening}
\title{Count the occurrence of a specific value in a column}
\usage{
count_screening(df)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data}
}
\value{
This function returns de-identified data.
}
\description{
De-identification of the TIMCI research data
}
