% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_submissions.R
\name{count_submissions}
\alias{count_submissions}
\title{Count submissions (TIMCI-specific)}
\usage{
count_submissions(
  facility_data,
  raw_day7fu_data,
  raw_hospit_data,
  raw_day28fu_data
)
}
\arguments{
\item{facility_data}{Dataframe that contains TIMCI Day 0 facility data.}

\item{raw_day7fu_data}{Dataframe that contains TIMCI Day 7 follow-up data.}

\item{raw_hospit_data}{Dataframe that contains TIMCI hospital follow-up data.}

\item{raw_day28fu_data}{Dataframe that contains TIMCI Day 28 follow-up data.}
}
\description{
Count submissions (TIMCI-specific)
}
