% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_day28fu_outcome_qc_flowchart}
\alias{create_day28fu_outcome_qc_flowchart}
\title{Create cleaning flowchart for successful Day 28 follow-up data (TIMCI-specific)}
\usage{
create_day28fu_outcome_qc_flowchart(
  n_raw_day28fu_records,
  n_dropped_duplicate_records,
  n_fu_prior_day0_day28fu,
  n_death_prior_day0_day28fu,
  n_hospit_prior_day0_day28fu,
  n_death_prior_hospit_day28fu,
  n_cleaned_day28fu_records
)
}
\arguments{
\item{n_raw_day28fu_records}{Number of records}

\item{n_dropped_duplicate_records}{TBD}

\item{n_cleaned_day28fu_records}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for successful Day 28 follow-up data (TIMCI-specific)
}
