% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_day7fu_qc_flowchart}
\alias{create_day7fu_qc_flowchart}
\title{Create cleaning flowchart for Day 7 follow-up data (TIMCI-specific)}
\usage{
create_day7fu_qc_flowchart(
  n_raw_records,
  n_nonvalid_pids,
  n_clean_records,
  d,
  e,
  f,
  g
)
}
\arguments{
\item{n_raw_records}{Number of records}

\item{n_nonvalid_pids}{Number of records with a non-valid participan ID}

\item{n_clean_records}{TBD}

\item{d}{TBD}

\item{e}{TBD}

\item{f}{TBD}

\item{g}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for Day 7 follow-up data (TIMCI-specific)
}
