% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_drug_qc_flowchart}
\alias{create_drug_qc_flowchart}
\title{Create cleaning flowchart for drug re-entry data (TIMCI-specific)}
\usage{
create_drug_qc_flowchart(
  n_raw_drug_records,
  n_dropped_duplicate_records,
  n_cleaned_drug_records
)
}
\arguments{
\item{n_raw_drug_records}{Initial number of drug records}

\item{n_dropped_duplicate_records}{Number of drug records that were dropped}

\item{n_cleaned_drug_records}{Number of cleaned drug records}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for drug re-entry data (TIMCI-specific)
}
