% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_odkapi.R
\name{create_odkc_draft_form}
\alias{create_odkc_draft_form}
\title{Create an ODK Central draft form
See reference ODK API documentation in \url{https://odkcentral.docs.apiary.io/#reference/forms/draft-form/creating-a-draft-form} for more details.}
\usage{
create_odkc_draft_form(un, pw, form_url)
}
\arguments{
\item{un}{string containing the ODK Central username.}

\item{pw}{string containing the password for user `un`.}

\item{form_url}{string containing the URL of the ODK Central form to be published, e.g. "https://sandbox.central.getodk.org/v1/projects/project-id/forms/form-id".}
}
\value{
This function returns the content of the request to the ODK Central API.
}
\description{
Create an ODK Central draft form
See reference ODK API documentation in \url{https://odkcentral.docs.apiary.io/#reference/forms/draft-form/creating-a-draft-form} for more details.
}
