% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_screening_qc_flowchart}
\alias{create_screening_qc_flowchart}
\title{Create cleaning flowchart for screening data (TIMCI-specific)}
\usage{
create_screening_qc_flowchart(
  n_raw_screening_records,
  n_nonvalid_deviceid_records,
  n_other_fid_records,
  n_before_startdate_records,
  n_after_lockdate_records,
  n_ineligible_cg_records,
  n_edited_repeat_visit_records,
  n_incorrect_date_setup_records,
  n_cleaned_screening_records
)
}
\arguments{
\item{n_raw_screening_records}{Initial number of screening records}

\item{n_nonvalid_deviceid_records}{Number of screening records with a non-valid device ID}

\item{n_other_fid_records}{Number of screening records corresponding to facility involved in another TIMCI study (India-specific)}

\item{n_before_startdate_records}{Number of screening records with an entry date anterior to the start date}

\item{n_after_lockdate_records}{Number of screening records with an entry date posterior to the lock date}

\item{n_ineligible_cg_records}{Number of screening records with an ineligible caregiver}

\item{n_edited_repeat_visit_records}{Number of screening records that were edited manually}

\item{n_incorrect_date_setup_records}{Number of screening records with an incorrect date that had to be edited manually}

\item{n_cleaned_screening_records}{Number of cleaned screening records}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for screening data (TIMCI-specific)
}
