% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_data.R
\name{csv_download_server}
\alias{csv_download_server}
\title{CSV data download module server-side processing}
\usage{
csv_download_server(input, output, session, vars)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{vars}{}
}
\value{
list with following components
\describe{
  \item{form}{reactive character string indicating x variable selection}
}
}
\description{
CSV data download module server-side processing
}
