% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_csv_data_module.R
\name{csv_download_ui}
\alias{csv_download_ui}
\title{CSV data download module user interface}
\usage{
csv_download_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
CSV data download module user interface
}
