% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_odk_data_table_module.R
\name{data_table_server}
\alias{data_table_server}
\title{Data table module server-side processing}
\usage{
data_table_server(id, df)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{df}{parameter containing a dataframe}
}
\description{
This module displays the ODK data table corresponding to the form selected by the user.
}
