% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data_table_module.R
\name{data_table_ui}
\alias{data_table_ui}
\title{Data table module user interface}
\usage{
data_table_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Data table module user interface
}
