% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{database_export}
\alias{database_export}
\title{Write dataframe to an Excel file, with a filename consisting of a prefix and a date stamp}
\usage{
database_export(df, idx, label, cdir, description)
}
\arguments{
\item{df}{quality check dataframe.}

\item{idx}{directory where the Excel file will be created.}

\item{label}{filename prefix}

\item{cdir}{Row names}

\item{description}{Row names}
}
\value{
creation timestamp of the Excel file
}
\description{
Write dataframe to an Excel file, with a filename consisting of a prefix and a date stamp
}
