% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataset_export}
\alias{dataset_export}
\title{Export datasets}
\usage{
dataset_export(df, idx, label, cdir, description)
}
\arguments{
\item{df}{quality check dataframe.}

\item{idx}{directory where the Excel file will be created.}

\item{label}{filename prefix}

\item{cdir}{Row names}

\item{description}{Row names}
}
\value{
creation timestamp of the Excel file
}
\description{
Export datasets
}
