% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{deidentify_data}
\alias{deidentify_data}
\title{Deidentify data}
\usage{
deidentify_data(df)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data}
}
\value{
This function returns de-identified data.
}
\description{
De-identification of the TIMCI research data
}
