% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify_spa_data}
\alias{deidentify_spa_data}
\title{De-identify SPA data (TIMCI-specific Function)}
\usage{
deidentify_spa_data(df, mapping = NULL)
}
\arguments{
\item{df}{A dataframe containing the processed facility data. This should include a column for child identification numbers that will be anonymized.}

\item{mapping}{Optional. A dataframe containing the mapping between original and hashed IDs. If provided, it will be used for anonymization; otherwise, IDs are anonymized without a predefined mapping.}
}
\value{
Returns a dataframe with de-identified data. The 'child_identification_pid' column is anonymized to protect patient privacy.
}
\description{
This function performs de-identification of the TIMCI SPA observation data.
}
