% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify_spa_rx_data}
\alias{deidentify_spa_rx_data}
\title{De-identify SPA treatment data (TIMCI-specific Function)}
\usage{
deidentify_spa_rx_data(df, mapping = NULL)
}
\arguments{
\item{df}{A dataframe containing the processed treatment data, which includes a 'PARENT_KEY' column for treatment identifiers that need to be anonymized.}

\item{mapping}{Optional. A dataframe containing the mapping between original and hashed IDs. If provided, it uses this mapping for anonymization; otherwise, the function anonymizes IDs without a predefined mapping.}
}
\value{
Returns a dataframe with de-identified treatment data. The 'PARENT_KEY' column is anonymized to protect individual treatment records' privacy.
}
\description{
This function de-identifies treatment data within the TIMCI research dataset, focusing on anonymizing treatment identifiers to ensure privacy.
}
