% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{delete_day0_child_ids}
\alias{delete_day0_child_ids}
\title{Drop incorrect child IDs in Day 0 data entries (TIMCI-specific function)
This function can be used to drop documented child IDs}
\usage{
delete_day0_child_ids(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains a cleaned dataframe and the list of dropped records
}
\description{
Drop incorrect child IDs in Day 0 data entries (TIMCI-specific function)
This function can be used to drop documented child IDs
}
