% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_info.R
\name{detect_blank_var}
\alias{detect_blank_var}
\title{Detect blank ("") value (TIMCI-specific function)}
\usage{
detect_blank_var(df, col)
}
\arguments{
\item{df}{dataframe}

\item{col}{column name to check for missingness}
}
\value{
This function returns a dataframe containing only participants with no information in column `col`
}
\description{
Detect blank ("") value (TIMCI-specific function)
}
