% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drugs.R
\name{detect_drug}
\alias{detect_drug}
\title{Detect specific string in free text entries where structured entry is missing (TIMCI-specific function)}
\usage{
detect_drug(df, col_val, col_text, vec)
}
\arguments{
\item{df}{dataframe}

\item{col_val}{column name}

\item{col_text}{column name}

\item{vec}{vector of text to test}
}
\value{
This function returns a dataframe containing only participants with no information in column `col`
}
\description{
Detect specific string in free text entries where structured entry is missing (TIMCI-specific function)
}
