% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_duplicate_participants}
\alias{detect_duplicate_participants}
\title{Detect duplicate participants (TIMCI-specific function)}
\usage{
detect_duplicate_participants(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of possible duplicate participants only
}
\description{
Detect duplicate participants (TIMCI-specific function)
}
