% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_id_duplicates}
\alias{detect_id_duplicates}
\title{Detect ID duplicates (TIMCI-specific function)}
\usage{
detect_id_duplicates(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a vector containing IDs of duplicate IDs
}
\description{
Detect ID duplicates (TIMCI-specific function)
}
