% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_inconsistent_dates}
\alias{detect_inconsistent_dates}
\title{Detect inconsistent dates (TIMCI-specific function)}
\usage{
detect_inconsistent_dates(
  df,
  col_date_start,
  col_date_end,
  cleaning = "none",
  date_format = "\%Y-\%m-\%d \%T"
)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{col_date_start}{column}

\item{col_date_end}{column}

\item{cleaning}{type of cleaning to be performed on inconsistent dates, by default set to "none" (i.e., no cleaning following the identification of inconsistent dates)}

\item{date_format}{format of the date}
}
\value{
This function returns a dataframe containing data of possible duplicate participants only
}
\description{
Detect inconsistent dates (TIMCI-specific function)
}
