% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{detect_inconsistent_names_between_visits}
\alias{detect_inconsistent_names_between_visits}
\title{Detect inconsistent participant names between follow-up and baseline visits (TIMCI-specific function)}
\usage{
detect_inconsistent_names_between_visits(refdf, fudf)
}
\arguments{
\item{refdf}{reference dataframe}

\item{fudf}{dataframe containing the follow-up data to check}
}
\value{
This function returns a dataframe.
}
\description{
Detect inconsistent participant names between follow-up and baseline visits (TIMCI-specific function)
}
