% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_missing_clinical_presentation}
\alias{detect_missing_clinical_presentation}
\title{Detect missing clinical presentation (TIMCI-specific function)}
\usage{
detect_missing_clinical_presentation(facility_df)
}
\arguments{
\item{facility_df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing only participants with no clinical presentation
}
\description{
Detect missing clinical presentation (TIMCI-specific function)
}
