% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_info.R
\name{detect_missing_diagnosis}
\alias{detect_missing_diagnosis}
\title{Detect missing diagnosis (TIMCI-specific function)}
\usage{
detect_missing_diagnosis(facility_df)
}
\arguments{
\item{facility_df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing only participants with no diagnosis
}
\description{
Detect missing diagnosis (TIMCI-specific function)
}
