% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_info.R
\name{detect_missing_value}
\alias{detect_missing_value}
\title{Detect missing (NA) value (TIMCI-specific function)}
\usage{
detect_missing_value(df, col)
}
\arguments{
\item{df}{dataframe}

\item{col}{column name to check for missing (NA) values}
}
\value{
This function returns a dataframe containing only participants with no information in column `col`
}
\description{
Detect missing (NA) value (TIMCI-specific function)
}
