% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_info.R
\name{detect_missing_var}
\alias{detect_missing_var}
\title{Detect missing referral reported by caregiver at the exit of the consultation (TIMCI-specific function)}
\usage{
detect_missing_var(df, col)
}
\arguments{
\item{df}{dataframe}

\item{col}{column name to check for missingness}
}
\value{
This function returns a dataframe containing only participants with no information in column `col`
}
\description{
Detect missing referral reported by caregiver at the exit of the consultation (TIMCI-specific function)
}
