% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_name_duplicates}
\alias{detect_name_duplicates}
\title{Detect name duplicates (TIMCI-specific function)}
\usage{
detect_name_duplicates(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a vector containing IDs and names of duplicate names
}
\description{
Detect name duplicates (TIMCI-specific function)
}
