% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_namedob_duplicates}
\alias{detect_namedob_duplicates}
\title{Detect name and date of birth duplicates (TIMCI-specific function)}
\usage{
detect_namedob_duplicates(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a vector containing IDs of duplicate name
}
\description{
Detect name and date of birth duplicates (TIMCI-specific function)
}
