% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_info.R
\name{detect_negative_value}
\alias{detect_negative_value}
\title{Detect negative value (TIMCI-specific function)}
\usage{
detect_negative_value(df, col)
}
\arguments{
\item{df}{dataframe}

\item{col}{column name to check for negative values}
}
\value{
This function returns a dataframe containing only participants with no information in column `col`
}
\description{
Detect negative value (TIMCI-specific function)
}
