% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_non_timely_completion}
\alias{detect_non_timely_completion}
\title{Detect non-timely completion of a submission, i.e. submissions not finalised on the day they were started (ODK function)}
\usage{
detect_non_timely_completion(df)
}
\arguments{
\item{df}{dataframe containing any ODK data, assuming standard metadata fields (`start`, `end`) are present.}
}
\value{
This function returns a dataframe containing the duration between start and end
}
\description{
Detect non-timely completion of a submission, i.e. submissions not finalised on the day they were started (ODK function)
}
