% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_non_timely_submission}
\alias{detect_non_timely_submission}
\title{Detect non-timely submissions, i.e. submissions not sent to the server on the day they were finalised (ODK function)}
\usage{
detect_non_timely_submission(df)
}
\arguments{
\item{df}{dataframe containing any ODK data, assuming standard metadata fields (`start`, `end`) are present.}
}
\value{
This function returns a dataframe containing the duration between start and end
}
\description{
Detect non-timely submissions, i.e. submissions not sent to the server on the day they were finalised (ODK function)
}
