% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_participant_duplicates}
\alias{detect_participant_duplicates}
\title{Detect participant duplicates (TIMCI-specific function)
Detection based on name, sex and date of birth}
\usage{
detect_participant_duplicates(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a vector containing IDs of duplicate name
}
\description{
Detect participant duplicates (TIMCI-specific function)
Detection based on name, sex and date of birth
}
