% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{display_weekly_fa_data_per_facility}
\alias{display_weekly_fa_data_per_facility}
\title{Display weekly facility assessment data per facility (TIMCI-specific function)}
\usage{
display_weekly_fa_data_per_facility(df, all_df)
}
\arguments{
\item{df}{dataframe containing the processed ODK data collected in the weekly facility assessment}

\item{all_df}{dataframe containing the processed ODK data collected in the weekly facility assessment}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Display weekly facility assessment data per facility (TIMCI-specific function)
}
