% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_table_server.R
\name{dummy_data_table_server}
\alias{dummy_data_table_server}
\title{Module UI function}
\usage{
dummy_data_table_server(id, dummy_data)
}
\arguments{
\item{id}{Namespace for the Shiny module}

\item{dummy_data}{Dummy data}
}
\description{
Module UI function
}
