% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_table_ui.R
\name{dummy_data_table_ui}
\alias{dummy_data_table_ui}
\title{Module UI function}
\usage{
dummy_data_table_ui(id)
}
\arguments{
\item{id}{Namespace for the Shiny module}
}
\description{
Module UI function
}
