% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrolment_tab_server.R
\name{enrolment_tab_server}
\alias{enrolment_tab_server}
\title{Module UI function}
\usage{
enrolment_tab_server(id, data)
}
\arguments{
\item{id}{Namespace for the Shiny module}

\item{data}{Data}
}
\description{
Module UI function
}
