% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_df2rds}
\alias{export_df2rds}
\title{Write dataframe to an RDS file format, with a filename consisting of a prefix and a date stamp}
\usage{
export_df2rds(df, dirname, prefix)
}
\arguments{
\item{df}{dataframe.}

\item{dirname}{directory where the RDS file will be created.}

\item{prefix}{filename prefix}
}
\description{
Write dataframe to an RDS file format, with a filename consisting of a prefix and a date stamp
}
