% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_all_visits}
\alias{extract_all_visits}
\title{Extract visits (TIMCI-specific function)}
\usage{
extract_all_visits(df, is_pilot = FALSE)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{is_pilot}{Boolean, default set to `FALSE`}
}
\value{
This function returns a dataframe containing data of all baseline and repeat visits
}
\description{
Extract visits (TIMCI-specific function)
}
