% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{extract_baseline_visits}
\alias{extract_baseline_visits}
\title{Extract baseline visits (TIMCI-specific function)}
\usage{
extract_baseline_visits(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of baseline visits only
}
\description{
Extract baseline visits (TIMCI-specific function)
}
