% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_odk_data.R
\name{extract_data_from_odk_server}
\alias{extract_data_from_odk_server}
\title{Extract data from ODK server to a formatted dataframe.
This function relies on ruODK to export submissions.}
\usage{
extract_data_from_odk_server(
  cpid,
  cpid_forms,
  cfid,
  cpp = "",
  start_date = NULL,
  end_date = NULL,
  filter = NULL,
  col_specs = NULL,
  group = TRUE,
  split = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cpid}{integer, ODK project ID}

\item{cpid_forms}{list of form IDs in `cpid`}

\item{cfid}{string, ODK form ID}

\item{cpp}{string, ODK project passphrase (optional, only required if the ODK project is encrypted)}

\item{start_date}{date, data collection start date (optional, by default set to `NULL`)}

\item{end_date}{date, data collection end date (optional, by default set to `NULL`)}

\item{filter}{OData filter (optional, by default set to `NULL`)}

\item{col_specs}{column specifications (optional, by default set to `NULL`)}

\item{group}{keep group names (optional, by default set to `TRUE`)}

\item{split}{split “select multiple” choices into columns (optional, by default set to `FALSE`)}

\item{verbose}{boolean, displays more information about the function output}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Extract data from ODK server to a formatted dataframe.
This function relies on ruODK to export submissions.
}
