% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{extract_data_from_odk_zip}
\alias{extract_data_from_odk_zip}
\title{Extract data from ODK zip}
\usage{
extract_data_from_odk_zip(odk_zip)
}
\arguments{
\item{odk_zip}{The absolute path to the zip file named "`fid`.zip" containing ODK submissions as CSV, plus separate CSVs for any repeating groups, plus any attachments in a subfolder `media`}
}
\value{
This function returns a dataframe.
}
\description{
Extract data from ODK zip
}
