% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_enrolled_participants}
\alias{extract_enrolled_participants}
\title{Extract enrolled participants (TIMCI-specific function)}
\usage{
extract_enrolled_participants(df, is_pilot = FALSE)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{is_pilot}{Boolean, default set to `FALSE`}
}
\value{
This function returns a dataframe containing data of enrolled participants only
}
\description{
Extract enrolled participants (TIMCI-specific function)
}
