% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_hypoxaemia}
\alias{extract_hypoxaemia}
\title{Extract hypoxaemia (TIMCI-specific function)}
\usage{
extract_hypoxaemia(df)
}
\arguments{
\item{df}{Dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of hypoxemic children only
}
\description{
Extract hypoxaemia (TIMCI-specific function)
}
