% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_noneligible}
\alias{extract_noneligible}
\title{Extract non-enrolled participants (TIMCI-specific function)}
\usage{
extract_noneligible(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of enrolled participants only
}
\description{
Extract non-enrolled participants (TIMCI-specific function)
}
