% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_pii}
\alias{extract_pii}
\title{Extract personally identifiable information (TIMCI-specific function)}
\usage{
extract_pii(df, is_pilot = FALSE)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{is_pilot}{Boolean, default set to `FALSE`}
}
\value{
This function returns a list of 2 dataframes: 1 dataframe with pii and 1 dataframe with deidentified demographic data
}
\description{
Extract personally identifiable information (TIMCI-specific function)
}
