% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{extract_repeat_visits}
\alias{extract_repeat_visits}
\title{Extract repeat visits (TIMCI-specific function)}
\usage{
extract_repeat_visits(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of repeat visits only
}
\description{
Extract repeat visits (TIMCI-specific function)
}
