% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_screening_data}
\alias{extract_screening_data}
\title{Extract screening data (TIMCI-specific function)}
\usage{
extract_screening_data(df, is_pilot)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{is_pilot}{Boolean, default set to `FALSE`}
}
\value{
This function returns a dataframe containing screening data only
}
\description{
Extract screening data (TIMCI-specific function)
}
