% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_spatf_data.R
\name{extract_tf_data}
\alias{extract_tf_data}
\title{Extract Time-Flow Data (TIMCI-specific function)}
\usage{
extract_tf_data(df)
}
\arguments{
\item{df}{data frame containing the raw time-flow data}
}
\value{
A data frame containing time-flow data, as specified by the Time-Flow Dictionary.
}
\description{
This function takes a data frame containing raw time-flow data and extracts the relevant
columns according to the Time-Flow Dictionary. The Time-Flow Dictionary is obtained using
the \code{\link{read_tf_xls_dict}} function.
}
