% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{find_best_matched_names_between_fu_and_day0}
\alias{find_best_matched_names_between_fu_and_day0}
\title{Detects matched names between follow-up and day 0 dataframes (TIMCI-specific function).}
\usage{
find_best_matched_names_between_fu_and_day0(
  df,
  day0_df,
  col_date,
  col_name,
  ldate_diff = 0,
  udate_diff = 0
)
}
\arguments{
\item{df}{Follow-up dataframe to search for matched names.}

\item{day0_df}{Day 0 dataframe to search for matched names.}

\item{col_date}{Name of the column containing the date in the `df` dataframe.}

\item{col_name}{Name of the column containing the name in the `df` dataframe.}

\item{ldate_diff}{Lower date difference (default is same day), negative numbers indicate a difference in the past, positive numbers indicate a difference in the future.}

\item{udate_diff}{Upper date difference (default is same day), negative numbers indicate a difference in the past, positive numbers indicate a difference in the future.}
}
\value{
A list with two dataframes: qc_df and cleaned_df.
}
\description{
This function searches for matched names between a follow-up dataframe and a day 0 dataframe. It returns a list containing two dataframes: one with quality control information, and another one with cleaned data.
}
