% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gps_coordinates.R
\name{find_closest_facility}
\alias{find_closest_facility}
\title{Find Closest Facility}
\usage{
find_closest_facility(df, facilities)
}
\arguments{
\item{df}{A data frame containing longitude and latitude information for each point}

\item{facilities}{A data frame containing longitude and latitude information for each facility}
}
\value{
A data frame with the closest facilities within a distance of 100 units
}
\description{
This function finds the closest facility from a data frame to a set of facilities
}
