% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fu_data.R
\name{format_day28_data}
\alias{format_day28_data}
\title{Process day 28 follow-up data (TIMCI-specific function)}
\usage{
format_day28_data(df, is_deidentified = TRUE)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data collected during the Day 28 follow-up call}

\item{is_deidentified}{Boolean, flag to not export personally identifiable variable (default set to `TRUE`)}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process day 28 follow-up data (TIMCI-specific function)
}
