% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fu_data.R
\name{format_day7_data}
\alias{format_day7_data}
\title{Process day 7 follow-up data (TIMCI-specific function)}
\usage{
format_day7_data(df)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data collected during the Day 7 follow-up call}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process day 7 follow-up data (TIMCI-specific function)
}
