% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_odk_data.R
\name{format_multiselect_asws}
\alias{format_multiselect_asws}
\title{Format multiple select answers so as to separate them}
\usage{
format_multiselect_asws(df, cols, sep)
}
\arguments{
\item{df}{dataframe containing ODK data}

\item{cols}{list of column names}

\item{sep}{separator, e.g. ";" "," etc}
}
\value{
This function returns a dataframe with multiple answers separated by `sep`.
}
\description{
Format multiple select answers so as to separate them
}
