% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_utils.R
\name{format_nrow}
\alias{format_nrow}
\title{Format display number of rows}
\usage{
format_nrow(s_before, df, s_after)
}
\arguments{
\item{s_before}{String displayed before the number of rows}

\item{df}{Dataframe}

\item{s_after}{String displayed after the number of rows}
}
\description{
Format display number of rows
}
