% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{format_odk_data}
\alias{format_odk_data}
\title{Format ODK data}
\usage{
format_odk_data(data)
}
\arguments{
\item{data}{dataframe containing the current ODK data}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Description of format_odk_data
}
