% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{format_summary_table}
\alias{format_summary_table}
\title{Format summary table to have a global descriptive statistics on the first row and a descriptive statistics by group in subsequent rows}
\usage{
format_summary_table(all, stat_list, gvar)
}
\arguments{
\item{all}{Dataframe}

\item{stat_list}{List of lists of formulea for summarizing the dataset `all`}

\item{gvar}{Grouping variable}
}
\value{
This function returns a statistical summary table ready to be printed
}
\description{
Format summary table to have a global descriptive statistics on the first row and a descriptive statistics by group in subsequent rows
}
