% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_odk_data.R
\name{format_text_fields}
\alias{format_text_fields}
\title{Replace commas by semicolons in text fields for CSV export}
\usage{
format_text_fields(df, cols)
}
\arguments{
\item{df}{dataframe containing ODK data}

\item{cols}{list of column names}
}
\value{
This function returns a dataframe with multiple answers separated by `sep`.
}
\description{
Replace commas by semicolons in text fields for CSV export
}
