% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{format_weekly_fa_data_for_export}
\alias{format_weekly_fa_data_for_export}
\title{Format weekly facility assessment data for export (TIMCI-specific function)}
\usage{
format_weekly_fa_data_for_export(df)
}
\arguments{
\item{df}{dataframe containing the processed facility assessment data}
}
\value{
This function returns a formatted dataframe for export.
}
\description{
Format weekly facility assessment data for export (TIMCI-specific function)
}
