% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formats2h1}
\alias{formats2h1}
\title{Format string to header of level H1 for printing in console}
\usage{
formats2h1(s)
}
\arguments{
\item{s}{String}
}
\description{
This function formats an input string `s` for printing as a header of level H1 in an output console
}
