% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_utils.R
\name{formats2h3}
\alias{formats2h3}
\title{Format string to header of level H3 for printing in console}
\usage{
formats2h3(s)
}
\arguments{
\item{s}{String to format}
}
\description{
This function formats an input string `s` for printing as a header of level H2 in an output console
}
