% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_calendar_heatmap}
\alias{generate_calendar_heatmap}
\title{Generate a calendar heatmap}
\usage{
generate_calendar_heatmap(df, datecol, end_date = Sys.Date())
}
\arguments{
\item{df}{Data frame to use for the plot}

\item{datecol}{Column name in data frame `df` that contains dates}

\item{end_date}{Final date for drawing the heatmap calendar (optional)}
}
\value{
This function returns a ggplot object which contains a calendar heatmap.
}
\description{
Generate a calendar heatmap
}
