% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{generate_calendar_heatmap2}
\alias{generate_calendar_heatmap2}
\title{Generate a calendar heatmap}
\usage{
generate_calendar_heatmap2(
  df,
  datecol,
  legendtitle = "Values",
  date_max = Sys.Date(),
  date_min = NULL
)
}
\arguments{
\item{df}{Data frame to use for the plot}

\item{datecol}{Column name in data frame `df` that contains dates}

\item{legendtitle}{Legend title (optional).}

\item{date_max}{End date of the plot (optional).}

\item{date_min}{Start date of the plot (optional).}
}
\value{
This function returns a ggplot object which contains a calendar heatmap.
}
\description{
Generate a calendar heatmap
}
