% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{generate_cg_log}
\alias{generate_cg_log}
\title{Generate caregiver recruitment log}
\usage{
generate_cg_log(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of caregiver to be called for the qualitative studies
}
