% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{generate_day7_fu_log}
\alias{generate_day7_fu_log}
\title{Generate follow-up log}
\usage{
generate_day7_fu_log(df, wmin, wmax)
}
\arguments{
\item{df}{dataframe}

\item{wmin}{dataframe}

\item{wmax}{dataframe}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants to be called at Day between wmin and wmax
}
