% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_day_bar_plot}
\alias{generate_day_bar_plot}
\title{Create a daily bar plot}
\usage{
generate_day_bar_plot(df, date_min, date_max)
}
\arguments{
\item{df}{Dataframe to use for the plot}

\item{date_min}{Start of the plot}

\item{date_max}{End of the plot}
}
\value{
This function returns a ggplot object which contains a bar plot.
}
\description{
generate_day_bar_plot() creates a bar plot.
}
