% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_day_cumbar_plot}
\alias{generate_day_cumbar_plot}
\title{Create a cumulative daily bar plot}
\usage{
generate_day_cumbar_plot(
  date_vec_list,
  lbl_vec,
  date_min,
  date_max,
  ylim = NULL,
  ylbl = "Frequencies"
)
}
\arguments{
\item{date_vec_list}{List of vectors containing dates}

\item{lbl_vec}{Vector of strings containing the name of each group}

\item{date_min}{Start date of the plot}

\item{date_max}{End date of the plot}

\item{ylim}{Numeric vector of length 2 that contains the min and max values of the y-axis}

\item{ylbl}{String of the y-axis}
}
\value{
This function returns a ggplot object which contains a cumulative bar plot of frequencies by dates
}
\description{
generate_day_cumbar_plot() creates a bar plot.
}
