% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{generate_dh_heatmap}
\alias{generate_dh_heatmap}
\title{Create a heatmap by day of Week and hour of day}
\usage{
generate_dh_heatmap(df, datecol)
}
\arguments{
\item{df}{Data frame to use for the plot}

\item{datecol}{Column name in data frame `df` that contains dates}
}
\value{
This function returns a ggplot object which contains a heatmap by day of Week and hour of day.
}
\description{
Create a heatmap by day of Week and hour of day
}
