% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_enrolment_hist}
\alias{generate_enrolment_hist}
\title{Create an enrolment histogram}
\usage{
generate_enrolment_hist(df)
}
\arguments{
\item{df}{Dataframe to use for the plot}
}
\value{
This function returns a ggplot object which contains a histogram representing the enrolment in each facility.
}
\description{
generate_enrolment_hist() creates a histogram plot that represents the enrolment level in the different health facilities.
Red bars indicate health facilities which are far below the enrolment target.
Orange bars indicate health facilities which are slightly below the enrolment target.
Green bars indicate health facilities which are above the enrolment target.
}
