% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_folder_structure.R
\name{generate_folder_structure}
\alias{generate_folder_structure}
\title{Generate folder structure (TIMCI-specific function)}
\usage{
generate_folder_structure(output_dir, study_dirname)
}
\arguments{
\item{output_dir}{directory}

\item{study_dirname}{directory}
}
\value{
This function returns a list of directories
}
\description{
Generate folder structure (TIMCI-specific function)
}
