% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{generate_fu_log}
\alias{generate_fu_log}
\title{Generate follow-up log (TIMCI-specific function)}
\usage{
generate_fu_log(df, fudf, wmin, wmax)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{fudf}{dataframe containing the processed follow-up data}

\item{wmin}{numerical, start of the follow-up period (in days)}

\item{wmax}{numerical, end of the follow-up period (in days)}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants to be called in a time window after baseline between wmin and wmax
}
