% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{generate_fu_log2}
\alias{generate_fu_log2}
\title{Generate follow-up log 2 (TIMCI-specific function)}
\usage{
generate_fu_log2(pii, fudf, wmin, wmax, vwmin, vwmax)
}
\arguments{
\item{pii}{dataframe containing personally identifiable data}

\item{fudf}{dataframe containing the processed follow-up data}

\item{wmin}{numerical, start of the follow-up period (in days)}

\item{wmax}{numerical, end of the follow-up period (in days)}

\item{vwmin}{numerical, start of the follow-up period valid for the statistical analysis (in days)}

\item{vwmax}{numerical, end of the follow-up period valid for the statistical analysis (in days)}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants to be called in a time window after baseline between wmin and wmax
}
