% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fu_logs.R
\name{generate_fu_logs}
\alias{generate_fu_logs}
\title{General follow-up logs}
\usage{
generate_fu_logs(
  rctls_pid,
  rctls_pp,
  research_facilities,
  fu_dir,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{rctls_pid}{numeric ID of the RCT/LS ODK Central project (mandatory parameter)}

\item{rctls_pp}{passphrase (mandatory parameter)}

\item{research_facilities}{dataframe that contains the research facilities (mandatory parameter)}

\item{fu_dir}{path to the output folder for the follow-up exports (mandatory parameter)}

\item{start_date}{start date (optional parameter)}

\item{end_date}{end date (optional parameter)}
}
\description{
This function runs follow-up Rmarkdown files to generate standardised follow-up logs for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
