% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_hospitfu_data.R
\name{generate_hospital_log}
\alias{generate_hospital_log}
\title{Generate hospital follow-up log (TIMCI-specific function)}
\usage{
generate_hospital_log(pii, fu7df, hospitdf, deidentify = FALSE)
}
\arguments{
\item{pii}{dataframe containing personally identifiable data}

\item{fu7df}{dataframe containing day 7 follow-up data}

\item{hospitdf}{dataframe containing hospital follow-up data}

\item{deidentify}{boolean, enable the generation of a format with personally identifiable information (PII) for upload on ODK Central (if set to FALSE) or of a deidentified dataframe (if set to TRUE)}
}
\value{
This function returns a dataframe that contains the list of participants to be searched for at hospital level.
}
\description{
Generate a list of participants to be searched for at hospital level
}
