% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fu_data.R
\name{generate_ltfu_log}
\alias{generate_ltfu_log}
\title{Generate lost to follow-up log (TIMCI-specific function)}
\usage{
generate_ltfu_log(df, fudf, end_date, raw = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{fudf}{dataframe containing the processed follow-up data}

\item{end_date}{numerical, end of the follow-up period (in days)}

\item{raw}{boolean flag for selecting raw (TRUE) or processed (FALSE) data}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants who are lost to follow-up after wmax
}
