% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pdf_report.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Generate PDF report}
\usage{
generate_pdf_report(file, plot_list)
}
\arguments{
\item{file}{File name}

\item{plot_list}{List of plots}
}
\value{
This function generates a monitoring report.
}
\description{
This function generates a monitoring report.
}
\details{
Description of generate_pdf_report
}
