% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_pie_chart}
\alias{generate_pie_chart}
\title{Generate pie chart with label}
\usage{
generate_pie_chart(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
This function returns a ggplot object.
}
\description{
This function plots a pie chart.
}
