% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{generate_report}
\alias{generate_report}
\title{Generate Rmarkdown file}
\usage{
generate_report(export_dir, rmd_fn, output_fn)
}
\arguments{
\item{export_dir}{Path to the output folder for the generated Rmarkdown reports and database exports}

\item{rmd_fn}{Filename of the Rmarkdown file}

\item{output_fn}{Filename of the Rmarkdown rendered report}
}
\description{
This function generate a standardised Rmarkdown report in two formats: html and docx.
}
