% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_research_cdsa_report.R
\name{generate_research_cdsa_report}
\alias{generate_research_cdsa_report}
\title{Generate a report comparing research and CDSA data}
\usage{
generate_research_cdsa_report(
  research_facilities,
  report_dir,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{research_facilities}{Dataframe that contains the research facilities}

\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{start_date}{Qualitative data collection start date (optional)}

\item{end_date}{Qualitative data collection end date (optional)}
}
\description{
Generate a report comparing research and CDSA data
}
