% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_time_hist_plot}
\alias{generate_time_hist_plot}
\title{Generate an histogram of day times}
\usage{
generate_time_hist_plot(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
This function returns an histogram of the distribution of data over day times
}
\description{
Generate an histogram of day times
}
