% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_week_bar_plot2}
\alias{generate_week_bar_plot2}
\title{Create a weekly bar plot by adding values from a specific column}
\usage{
generate_week_bar_plot2(
  date_vec,
  values,
  date_min,
  date_max,
  ylim = NULL,
  ylbl = "Values"
)
}
\arguments{
\item{date_vec}{Vector containing dates}

\item{values}{Vector containing values to sum}

\item{date_min}{Start date of the plot}

\item{date_max}{End date of the plot}

\item{ylim}{Numeric vector of length 2 that contains the min and max values of the y-axis}

\item{ylbl}{String of the y-axis}
}
\value{
This function returns a ggplot object which contains a bar plot of frequencies by week
}
\description{
generate_week_bar_plot() creates an absolute or a relative bar plot.
}
