% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{generate_word_report}
\alias{generate_word_report}
\title{Generate a Microsoft Word report using Rmarkdown}
\usage{
generate_word_report(report_dir, rmd_fn, report_fn, rmd_params = list())
}
\arguments{
\item{report_dir}{Path to the folder where the generated DOCX Rmarkdown report is stored}

\item{rmd_fn}{Filename of the Rmarkdown file}

\item{report_fn}{Filename of the Rmarkdown rendered report}

\item{rmd_params}{List of parameters}
}
\description{
This function generates a standardised Rmarkdown report in a DOCX format
}
