% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{get_summary_by_facility}
\alias{get_summary_by_facility}
\title{Get summary statistics grouped by facility (TIMCI-specific function)}
\usage{
get_summary_by_facility(df)
}
\arguments{
\item{df}{Dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing summary statistics grouped by health facility
}
\description{
Get summary statistics grouped by facility (TIMCI-specific function)
}
