% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmap_wday_hourofday}
\alias{heatmap_wday_hourofday}
\title{Generates a heat map of the number of submissions per weekday and hour.}
\usage{
heatmap_wday_hourofday(df, date_col)
}
\arguments{
\item{df}{Data frame containing the ODK data that is to be used. Optional, defaults to NULL.}

\item{date_col}{String that specifies the date or time stamp column in the data which is to be examined.}
}
\value{
ggplot2 object
}
\description{
Takes any time stamp column in the data and aggregates by weekday and hour of the day. The generated plot then shows for each combination
the frequency.
}
