% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{identify_day0_duplicates}
\alias{identify_day0_duplicates}
\title{Identify repeat ID duplicates by dates (TIMCI-specific function)}
\usage{
identify_day0_duplicates(df, col_id, col_date)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{col_id}{name of the column containing IDs in `df`}

\item{col_date}{name of the column containing dates in `df`}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify repeat ID duplicates by dates (TIMCI-specific function)
}
