% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{identify_day0_duplicates_and_fu}
\alias{identify_day0_duplicates_and_fu}
\title{Identify day 0 ID duplicates wit names and corresponding Day 7 follow-ups (TIMCI-specific function)}
\usage{
identify_day0_duplicates_and_fu(df, day7fu_df = NULL, cleaning = "none")
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{day7fu_df}{dataframe containing Day 7 follow-up data}

\item{cleaning}{type of cleaning to be performed on duplicates, by default set to "none" (i.e., no cleaning following the identification of duplicates)}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify day 0 ID duplicates wit names and corresponding Day 7 follow-ups (TIMCI-specific function)
}
