% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_dayfu_id_duplicates_with_dates}
\alias{identify_dayfu_id_duplicates_with_dates}
\title{Identify follow-up ID duplicates with dates (TIMCI-specific function)}
\usage{
identify_dayfu_id_duplicates_with_dates(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify follow-up ID duplicates with dates (TIMCI-specific function)
}
