% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_duplicates_with_dates}
\alias{identify_duplicates_with_dates}
\title{Identify child ID duplicates with dates (TIMCI-specific function)}
\usage{
identify_duplicates_with_dates(df, date_col)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{date_col}{name of the column containing dates in `df`}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify child ID duplicates with dates (TIMCI-specific function)
}
