% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_ids_outside_lock_range}
\alias{identify_ids_outside_lock_range}
\title{Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)}
\usage{
identify_ids_outside_lock_range(df, col_id, day0_df, start_date, end_date)
}
\arguments{
\item{df}{dataframe containing the data to check}

\item{col_id}{column name containing IDs in `df1`}

\item{day0_df}{Day 0 dataframe}

\item{start_date}{start date}

\item{end_date}{end date}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)
}
