% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{identify_multiple_enrolments}
\alias{identify_multiple_enrolments}
\title{Identify multiple 28-disease episodes for the same children (TIMCI-specific function)}
\usage{
identify_multiple_enrolments(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe.
}
\description{
Identify multiple 28-disease episodes for the same children (TIMCI-specific function)
}
