% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_nonvalid_ids}
\alias{identify_nonvalid_ids}
\title{Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)}
\usage{
identify_nonvalid_ids(df1, idcol1, df2, idcol2)
}
\arguments{
\item{df1}{dataframe containing the data to check}

\item{idcol1}{column name containing IDs in `df1`}

\item{df2}{reference dataframe}

\item{idcol2}{column name containing IDs in `df2`}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)
}
