% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_nonvalid_ids2}
\alias{identify_nonvalid_ids2}
\title{Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)}
\usage{
identify_nonvalid_ids2(df1, col_id1, df2, col_id2, col_date, lock_date)
}
\arguments{
\item{df1}{dataframe containing the data to check}

\item{col_id1}{column name containing IDs in `df1`}

\item{df2}{reference dataframe}

\item{col_id2}{column name containing IDs in `df2`}

\item{col_date}{column name containing dates in `df1`}

\item{lock_date}{lock date}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)
}
