% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_nonvalid_ids_flagged}
\alias{identify_nonvalid_ids_flagged}
\title{Identify Non-Valid IDs in a Dataframe Based on Reference IDs (TIMCI-specific)}
\usage{
identify_nonvalid_ids_flagged(df1, col_id1, df2, col_id2)
}
\arguments{
\item{df1}{A dataframe containing the data to be checked.}

\item{col_id1}{The name of the column containing IDs in the \code{df1} dataframe.}

\item{df2}{A reference dataframe used for comparison.}

\item{col_id2}{The name of the column containing reference IDs in the \code{df2} dataframe.}
}
\value{
A list containing two dataframes:
  \item{Non-Valid IDs Dataframe}{A dataframe containing rows with non-valid IDs and related dates, grouped in relevant columns.}
  \item{Flagged Dataframe}{A dataframe similar to \code{df1} with an additional column indicating if each ID is flagged as non-valid (1) or not (0).}
}
\description{
This function identifies non-valid IDs in a dataframe by comparing them with reference IDs from another dataframe. It is specifically designed for TIMCI data quality checks.
}
