% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{identify_nonvalid_ids_with_previous_duplicates}
\alias{identify_nonvalid_ids_with_previous_duplicates}
\title{Identify non-valid IDs in a dataframe based on IDs in another dataframe (TIMCI-specific function)}
\usage{
identify_nonvalid_ids_with_previous_duplicates(
  df1,
  col_id1,
  df2,
  col_id2,
  df3,
  col_id3,
  col_val
)
}
\arguments{
\item{df1}{A dataframe containing the data to check for non-valid IDs.}

\item{col_id1}{The column name containing IDs in \code{df1}.}

\item{df2}{A reference dataframe containing the valid IDs to compare with.}

\item{col_id2}{The column name containing IDs in \code{df2}.}

\item{df3}{A dataframe containing additional information related to valid IDs.}

\item{col_id3}{The column name containing IDs in \code{df3}.}

\item{col_val}{The column name containing values related to valid IDs in \code{df3}.}
}
\value{
A list containing two data frames. The first data frame contains the IDs and dates at which the ID has been allocated in different columns. The second data frame contains the cleaned data.
}
\description{
This function takes in two data frames and two column names, and identifies the non-valid IDs in the first data frame based on the IDs in the second data frame. It returns a list of two data frames, one containing the IDs and dates at which the ID has been allocated, and the other containing the cleaned data.
}
