% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{identify_true_duplicate}
\alias{identify_true_duplicate}
\title{Identify true ID duplicates by dates (TIMCI-specific function)}
\usage{
identify_true_duplicate(df, col_id, col_date, cleaning = "none")
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{col_id}{name of the column containing IDs in `df`}

\item{col_date}{name of the column containing dates in `df`}

\item{cleaning}{type of cleaning to be performed on duplicates, by default set to "none" (i.e., no cleaning following the identification of duplicates)}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Identify true ID duplicates by dates (TIMCI-specific function)
}
