% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_col_specifications}
\alias{import_col_specifications}
\title{Import column-specifications from country-specific Excel dictionary}
\usage{
import_col_specifications(xls_dict, country)
}
\arguments{
\item{xls_dict}{Excel spreadsheet containing 2 columns ('old' and 'new') that map the names of the variables in the input dataframe and the names of the variables in the output dataframe}

\item{country}{Character that contains the name of the TIMCI country to select (default set to "none")}
}
\value{
This function returns a dataframe.
}
\description{
Import column-specifications from country-specific Excel dictionary
}
