% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_not_empty}
\alias{is_not_empty}
\title{Check that a datframe is not empty}
\usage{
is_not_empty(df)
}
\arguments{
\item{df}{dataframe.}
}
\value{
This function returns a boolean that is equal to TRUE if the dataframe contains at least 1 row and 1 column.
}
\description{
Check that a datframe is not empty
}
