% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dummy_data.R
\name{load_dummy_data}
\alias{load_dummy_data}
\title{Load dummy data}
\usage{
load_dummy_data()
}
\value{
This function processes ODK data.
}
\description{
This function loads dummy data.
}
\details{
Description of load_dummy_data
}
