% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timci_odk_data.R
\name{load_qual_data}
\alias{load_qual_data}
\title{Load RCT/LS data}
\usage{
load_qual_data(
  odkc_project_list,
  qpid,
  qual_pp,
  qual_start_date,
  qual_end_date
)
}
\arguments{
\item{odkc_project_list}{List of ODK project IDs on the ODK Central server}

\item{qpid}{Numeric value that refers to the ID of the qualitative project on the ODK Central server}

\item{qual_pp}{Encryption passphrase associated with the qualitative ODK Central project}

\item{qual_start_date}{Qualitative data collection start date}

\item{qual_end_date}{Qualitative data collection end date}
}
\value{
list of ODK central project IDs
}
\description{
This function loads RCT/LS data for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
