% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{load_qual_data_and_generate_report}
\alias{load_qual_data_and_generate_report}
\title{Load qualitative data and generate corresponding report}
\usage{
load_qual_data_and_generate_report(
  qpid,
  qual_pp,
  research_facilities,
  report_dir,
  qual_start_date = NULL,
  qual_end_date = NULL
)
}
\arguments{
\item{qpid}{Numeric ID of the qualitative ODK Central project}

\item{qual_pp}{Encryption passphrase associated with the qualitative ODK Central project}

\item{research_facilities}{Dataframe that contains the research facilities}

\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{qual_start_date}{Qualitative data collection start date (optional)}

\item{qual_end_date}{Qualitative data collection end date (optional)}
}
\description{
This function loads qualitative data and run the qualitative Rmarkdown files to generate standardised automated reports for the qualitative studies in the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
