% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timci_odk_data.R
\name{load_rctls_data}
\alias{load_rctls_data}
\title{Load RCT/LS data}
\usage{
load_rctls_data(
  odkc_project_list,
  rctls_pid,
  rctls_pp,
  rctls_start_date,
  rctls_end_date
)
}
\arguments{
\item{odkc_project_list}{List of ODK project IDs on the ODK Central server}

\item{rctls_pid}{Numeric value that refers to the ID of the RCT/LS project on the ODK Central server}

\item{rctls_pp}{Encryption passphrase associated with the RCT/LS ODK ODK Central project}

\item{rctls_start_date}{RCT/LS data collection start date}

\item{rctls_end_date}{RCT/LS data collection end date}
}
\value{
list of ODK central project IDs
}
\description{
This function loads RCT/LS data for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
