% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timci_odk_data.R
\name{load_spapmtf_data}
\alias{load_spapmtf_data}
\title{Load Service Provision Assessment (SPA), process mapping & time-flow (PM&TF) data}
\usage{
load_spapmtf_data(odkc_project_list, spa_pid, spa_start_date, spa_end_date)
}
\arguments{
\item{odkc_project_list}{List of ODK project IDs on the ODK Central server}

\item{spa_pid}{Numeric value that refers to the ID of the SPA / time-flow / process mapping project on the ODK Central server}

\item{spa_start_date}{SPA data collection start date}

\item{spa_end_date}{SPA data collection end date}
}
\value{
list of ODK central project IDs
}
\description{
This function loads SPA data for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
