% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_from_dict}
\alias{match_from_dict}
\title{Extract and match variable names using a dictionary}
\usage{
match_from_dict(df, dictionary)
}
\arguments{
\item{df}{Input dataframe}

\item{dictionary}{Dataframe containing 2 columns ('old' and 'new') that map the names of the variables in the input dataframe and the names of the variables in the output dataframe}
}
\value{
This function returns a dataframe.
}
\description{
Extract and match variable names using a dictionary
}
