% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{match_from_drug_xls_dict}
\alias{match_from_drug_xls_dict}
\title{Match facility data using the Drug dictionary adapted for each country to account for differences in the data collection (TIMCI-specific function)}
\usage{
match_from_drug_xls_dict(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a dataframe with columns that match the specified country dictionary.
}
\description{
Match facility data using the Drug dictionary adapted for each country to account for differences in the data collection (TIMCI-specific function)
}
