% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_from_filtered_xls_dict}
\alias{match_from_filtered_xls_dict}
\title{Extract and match variable names using an external Excel dictionary}
\usage{
match_from_filtered_xls_dict(
  df,
  xls_dict,
  is_deidentified = FALSE,
  country = "none"
)
}
\arguments{
\item{df}{Input dataframe}

\item{xls_dict}{Excel file containing 2 columns ('old' and 'new') that map the names of the variables in the input dataframe and the names of the variables in the output dataframe}

\item{is_deidentified}{Boolean, flag to not export personally identifiable variable (default set to FALSE)}

\item{country}{Character that contains the name of the TIMCI country to select (default set to "none")}
}
\value{
This function returns a dataframe.
}
\description{
Extract and match variable names using an external Excel dictionary
}
