% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalised_levenshtein_ratio}
\alias{normalised_levenshtein_ratio}
\title{Calculate a normalised Levenshtein ratio between two strings}
\usage{
normalised_levenshtein_ratio(s1, s2)
}
\arguments{
\item{s1}{string}

\item{s2}{string}
}
\value{
This function returns a normalised levenshtein ratio.
}
\description{
Calculate a normalised Levenshtein ratio between two strings
}
