% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_odk_data_info.R
\name{odk_data_info_server}
\alias{odk_data_info_server}
\title{ODK data info module server-side processing}
\usage{
odk_data_info_server(input, output, session, vars)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{vars}{}
}
\value{
list with following components
\describe{
  \item{form}{reactive character string indicating x variable selection}
}
}
\description{
ODK data info module server-side processing
}
