% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_odk_data_info.R
\name{odk_data_info_ui}
\alias{odk_data_info_ui}
\title{ODK data info module user interface}
\usage{
odk_data_info_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
ODK data info module user interface
}
