% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_odk_data_table.R
\name{odk_data_table_server}
\alias{odk_data_table_server}
\title{ODK data table module server-side processing}
\usage{
odk_data_table_server(input, output, session, vars)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{vars}{parameter containing reactive list (called `vars`) for selecting the ODK data table to be displayed}
}
\description{
This module displays the ODK data table for the form selected by the user.
}
