% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_odk_data_table.R
\name{odk_data_table_ui}
\alias{odk_data_table_ui}
\title{ODK data table module user interface}
\usage{
odk_data_table_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
ODK data table module user interface
}
