% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart_module.R
\name{pie_chart_server}
\alias{pie_chart_server}
\title{Pie chart module server-side processing}
\usage{
pie_chart_server(id, data)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{data}{parameter containing a vector}
}
\description{
Pie chart module server-side processing
}
