% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{pie_chart_ui}
\alias{pie_chart_ui}
\title{Pie chart module user interface}
\usage{
pie_chart_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
Pie chart module user interface
}
