% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_duplicates_to_columns}
\alias{pivot_duplicates_to_columns}
\title{Pivot values of duplicated rows into new columns}
\usage{
pivot_duplicates_to_columns(df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
This function returns a dataframe.
}
\description{
Pivot values of duplicated rows into new columns
}
