% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enrolment_gauge}
\alias{plot_enrolment_gauge}
\title{Create an enrolment gauge}
\usage{
plot_enrolment_gauge(val, lbl, m, lthres, uthres)
}
\arguments{
\item{val}{Value}

\item{lbl}{Label}

\item{m}{Maximal gauge value}

\item{lthres}{Numeric value,  lower threshold}

\item{uthres}{Numeric value, upper threshold}
}
\value{
This function plots a gauge representing enrolment.
}
\description{
plot_enrolment_gauge() creates a gauge plot that represents the global enrolment with respect to the target.
}
\details{
Description of plot_enrolment_gauge
}
