% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gauge}
\alias{plot_gauge}
\title{Create an enrolment gauge}
\usage{
plot_gauge(val, lbl, m, lthres, uthres, scale = 1)
}
\arguments{
\item{val}{Value}

\item{lbl}{Label}

\item{m}{Maximal gauge value}

\item{lthres}{Numeric value,  lower threshold}

\item{uthres}{Numeric value, upper threshold}

\item{scale}{Numeric value (optional, default 1), must vary between 0 and 1 to scale the text size}
}
\value{
This function plots a gauge with rate and title.
}
\description{
plot_gauge() creates a gauge plot that represents the current rate with respect to a target.
}
\details{
Description of plot_enrolment_gauge
}
