% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_geom_bar_by_facility_over_time}
\alias{plot_geom_bar_by_facility_over_time}
\title{Create facets with bar plots over time}
\usage{
plot_geom_bar_by_facility_over_time(
  df,
  facility_col,
  date_col,
  date_lbl,
  date_break_str,
  date_format,
  start_date,
  end_date,
  max_val = NULL,
  fill_col = "",
  n_facet_per_row = 5,
  text_size = 7
)
}
\arguments{
\item{df}{Data frame to use for the plot.}

\item{facility_col}{Column name in data frame `df`.}

\item{date_col}{Column name in data frame `df`, should contain dates.}

\item{date_lbl}{String that contains the x-axis label for the plot.}

\item{date_break_str}{A string giving the distance between date breaks like "2 weeks", or "10 years".}

\item{date_format}{A string giving the date formatting specification for the date labels (for instance `\%b\%y`).}

\item{start_date}{Minimal value in the date range}

\item{end_date}{Maximal value in the date range}

\item{max_val}{Maximal limit for the y-axis (optional, default `NULL`)}

\item{fill_col}{Column name in data frame `df`(optional, default empty string "").}

\item{n_facet_per_row}{Numeric value (optional, default 5) to set the number of facets per rows.}

\item{text_size}{Numeric value (optional, default 7) to scale the text size.}
}
\value{
This function returns a ggplot2 object.
}
\description{
plot_geom_bar_by_facility_over_time() creates an image that contains facets with bar plots over time.
}
