% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_geom_bar_over_time}
\alias{plot_geom_bar_over_time}
\title{Plot Bar Chart over Time}
\usage{
plot_geom_bar_over_time(
  df,
  facility_col,
  date_col,
  date_lbl,
  date_break_str,
  date_format,
  start_date,
  end_date,
  max_val = NULL,
  text_size = 7
)
}
\arguments{
\item{df}{a dataframe containing the data to plot}

\item{facility_col}{a string indicating the column name of the facility variable in \code{df}}

\item{date_col}{a string indicating the column name of the date variable in \code{df}}

\item{date_lbl}{a string indicating the x-axis label for the plot}

\item{date_break_str}{a string indicating the date format for the x-axis ticks}

\item{date_format}{a string indicating the date format for the x-axis labels}

\item{start_date}{a string indicating the start date for the x-axis}

\item{end_date}{a string indicating the end date for the x-axis}

\item{max_val}{an optional numeric indicating the maximum y-value for the plot}

\item{text_size}{an optional numeric indicating the size of the text in the plot}
}
\value{
a ggplot object displaying the bar chart over time
}
\description{
This function creates a bar chart by area type over time, based on the data
provided in the input dataframe. The bars represent the number of records in
each facility for each time interval.
}
