% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_numeric_indicator}
\alias{plot_numeric_indicator}
\title{Create a numeric indicator}
\usage{
plot_numeric_indicator(val, lbl, scale = 1)
}
\arguments{
\item{val}{Value}

\item{lbl}{Label}

\item{scale}{Numeric value (optional, default 1), must vary between 0 and 1 to scale the text size}
}
\value{
This function returns a ggplot2 object.
}
\description{
plot_numeric_indicator() creates an image that contains a numeric indicator as well as a short description of the variable it represents.
}
