% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter_by_facility_over_time}
\alias{plot_scatter_by_facility_over_time}
\title{Create facets with scatters over time}
\usage{
plot_scatter_by_facility_over_time(
  df,
  facility_col,
  date_col,
  date_lbl,
  date_break_str,
  date_format,
  start_date,
  end_date,
  y_col,
  y_lbl,
  y_is_time = FALSE,
  time_break_str = "",
  time_format = "",
  y_min = NA,
  y_max = NA,
  n_facet_per_row = 5,
  text_size = 7
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{facility_col}{The name of the column in \code{df} that represents the facilities}

\item{date_col}{Column name in data frame `df`, should contain dates.}

\item{date_lbl}{String that contains the x-axis label for the plot.}

\item{date_break_str}{A string giving the distance between date breaks like "2 weeks", or "10 years".}

\item{date_format}{A string giving the date formatting specification for the date labels (for instance `\%b\%y`).}

\item{start_date}{Minimal value in the date range.}

\item{end_date}{Maximal value in the date range.}

\item{y_col}{Column name in data frame `df`.}

\item{y_lbl}{String that contains the y-axis label for the plot.}

\item{y_is_time}{Boolean (optional, default `FALSE`).}

\item{time_break_str}{A string giving the distance between time breaks like "4 hours" (optional, default "").}

\item{time_format}{A string giving the date formatting specification for the time labels, for instance `\%H:\%M` (optional, default "").}

\item{y_min}{Minimum value of the y-axis (optional, default `NA`).}

\item{y_max}{Maximum value of the y-axis (optional, default `NA`).}

\item{n_facet_per_row}{Numeric value (optional, default 5) to set the number of facets per rows.}

\item{text_size}{Numeric value (optional, default 7) to scale the text size.}
}
\value{
This function returns a ggplot2 object.
}
\description{
This function creates a scatter plot with multiple facets, where each facet represents a facility or location.
The plot is organized over time, with the x-axis representing dates and the y-axis representing values from a specified column in the data frame.
}
