% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{process_facility_data}
\alias{process_facility_data}
\title{Process facility data (TIMCI-specific function)}
\usage{
process_facility_data(df, is_pilot)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data collected at the facility level}

\item{is_pilot}{Boolean, default set to `FALSE`}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process facility data (TIMCI-specific function)
}
