% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_location_data.R
\name{process_gps_coordinates}
\alias{process_gps_coordinates}
\title{Process GPS coordinates (TIMCI-specific function)}
\usage{
process_gps_coordinates(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns cleaned GPS coordinates.
}
\description{
Process GPS coordinates (TIMCI-specific function)
}
