% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{process_odk_data}
\alias{process_odk_data}
\title{Process data from the ODK Central server}
\usage{
process_odk_data(data)
}
\arguments{
\item{data}{ODK data}
}
\value{
This function processes ODK data.
}
\description{
This function processes ODK data.
}
\details{
Description of get ODK data
}
