% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_spa_data.R
\name{process_spa_hcpi_data}
\alias{process_spa_hcpi_data}
\title{Process healthcare provider interview data (TIMCI-specific function)}
\usage{
process_spa_hcpi_data(df)
}
\arguments{
\item{df}{dataframe containing the (raw) ODK data collected in the healthcare provider interview}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process healthcare provider interview data (TIMCI-specific function)
}
