% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_spa_data.R
\name{process_spa_sco_data}
\alias{process_spa_sco_data}
\title{Process sick child observation data (TIMCI-specific function)}
\usage{
process_spa_sco_data(df)
}
\arguments{
\item{df}{dataframe containing the (raw) ODK data collected in the SPA sick child interview observation protocol form}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process sick child observation data (TIMCI-specific function)
}
