% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{process_weekly_fa_data}
\alias{process_weekly_fa_data}
\title{Process weekly facility assessment data (TIMCI-specific function)}
\usage{
process_weekly_fa_data(df)
}
\arguments{
\item{df}{dataframe containing the (raw) ODK data collected in the weekly facility assessment}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process weekly facility assessment data (TIMCI-specific function)
}
