% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_code_image_server.R
\name{qr_code_image_server}
\alias{qr_code_image_server}
\title{Module UI function}
\usage{
qr_code_image_server(id, outfile)
}
\arguments{
\item{id}{Namespace for the Shiny module}
}
\description{
Module UI function
}
