% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_modules.R
\name{reactive_csv_download_server}
\alias{reactive_csv_download_server}
\title{Reactive CSV data download module server-side processing}
\usage{
reactive_csv_download_server(id, vars)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{vars}{parameter containing a reactive list with the following components:
\describe{
  \item{`citem`}{string indicating the current ODK form}
  \item{`cdata`}{dataframe containing the current ODK data}
}}
}
\description{
Reactive CSV data download module server-side processing
}
