% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{read_day0_xls_dict}
\alias{read_day0_xls_dict}
\title{Load the Day 0 dictionary - adapted for each country to account for differences in the data collection (TIMCI-specific function)}
\usage{
read_day0_xls_dict()
}
\value{
This function returns a dataframe that will be used as a dictionary to convert data exported from ODK Central to a statistician-friendly format.
}
\description{
Load the Day 0 dictionary - adapted for each country to account for differences in the data collection (TIMCI-specific function)
}
