% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_spatf_data.R
\name{read_tf_xls_dict}
\alias{read_tf_xls_dict}
\title{Load the Time-Flow Dictionary Adapted for Each Country (TIMCI-specific function)}
\usage{
read_tf_xls_dict()
}
\value{
This function returns a dataframe that will be used as a dictionary to convert data exported from ODK Central to a statistician-friendly format.
}
\description{
This function imports a country-specific Excel dictionary, which is adapted for each
country to account for differences in data collection processes. The dictionary is
intended to be used for converting data exported from ODK Central into a format that is
more suitable for statistical analysis
}
