% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screening_facility.R
\name{remove_facilities_for_other_studies}
\alias{remove_facilities_for_other_studies}
\title{Remove facilities in which data collection is conducted for another study that the current study (India-specific)}
\usage{
remove_facilities_for_other_studies(
  facility_data,
  all_facilities,
  excluded_facilities
)
}
\arguments{
\item{facility_data}{dataframe containing the processed facility data}

\item{all_facilities}{dataframe containing the list of research facilities and their characteristics}

\item{excluded_facilities}{dataframe containing the list of research facilities to be excluded and their characteristics}
}
\description{
Remove facilities in which data collection is conducted for another study that the current study (India-specific)
}
