% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_modules.R
\name{report_download_server}
\alias{report_download_server}
\title{Report download module server-side processing}
\usage{
report_download_server(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
Report download module server-side processing
}
