% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_modules.R
\name{report_download_ui}
\alias{report_download_ui}
\title{Report download module user interface}
\usage{
report_download_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
Report download module user interface
}
