% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_data_table_ui.R
\name{research_data_table_ui}
\alias{research_data_table_ui}
\title{Module UI function}
\usage{
research_data_table_ui(id)
}
\arguments{
\item{id}{Namespace for the Shiny module}
}
\description{
Module UI function
}
