% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the TIMCI Shiny app}
\usage{
run_app()
}
\description{
This function runs a Shiny web application dedicated to the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
