% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mini_app}
\alias{run_mini_app}
\title{Run a minimalist Shiny app example}
\usage{
run_mini_app()
}
\description{
This function runs a minimalist Shiny web application with a simple connection to ODK Central.
}
