% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown}
\alias{run_rmarkdown}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown(export_dir)
}
\arguments{
\item{export_dir}{Path to the output folder for the generated Rmarkdown reports and database exports}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
