% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown_rctls}
\alias{run_rmarkdown_rctls}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown_rctls(
  rctls_pid,
  rctls_pp,
  research_facilities,
  report_dir,
  participant_zip,
  mdb_dir,
  path_dir,
  start_date = NULL,
  end_date = NULL,
  lock_date = NULL,
  sample_size = 1e+05,
  short = FALSE
)
}
\arguments{
\item{rctls_pid}{Numeric value that refers ID of the RCT/LS ODK Central project}

\item{rctls_pp}{Passphrase}

\item{research_facilities}{Dataframe that contains the research facilities}

\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{participant_zip}{Path to the encrypted zip archive that stores participant data}

\item{mdb_dir}{Path to the output folder for the RCT / LS database exports}

\item{path_dir}{Path to the output folder for the M&E exports to be shared with PATH}

\item{start_date}{RCT/LS data collection start date (optional)}

\item{end_date}{RCT/LS data collection end date (optional)}

\item{lock_date}{RCT/LS data collection cleaning end date for database lock (optional)}

\item{sample_size}{Numeric value, sample size for RCT/LS enrolment}

\item{short}{Short version of the export}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the longitudinal study (LS) and the randomised controlled trial (RCT) for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
