% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_list_item_module.R
\name{select_list_item_server}
\alias{select_list_item_server}
\title{List item selection module server-side processing}
\usage{
select_list_item_server(id, list, data)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{list}{list of strings (non-reactive)}

\item{data}{hash structure (non-reactive) indexing dataframes}
}
\value{
reactive list with the following components:
\describe{
  \item{`citem`}{string indicating the current ODK form}
  \item{`cdata`}{dataframe containing the current ODK data}
}
}
\description{
List item selection module server-side processing
}
