% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_list_item.R
\name{select_list_item_ui}
\alias{select_list_item_ui}
\title{List item selection module user interface}
\usage{
select_list_item_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
List item selection module user interface
}
