% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send email}
\usage{
send_email(report_dir, report_fn, dest)
}
\arguments{
\item{report_dir}{Path to the folder where the generated HTML and Microsoft Word DOCX Rmarkdown reports are stored}

\item{report_fn}{Filename of the Rmarkdown rendered report}

\item{dest}{vector of strings containing email addresses}
}
\description{
This function sends an automated email with the generated reports
}
