% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email_tls}
\alias{send_email_tls}
\title{Send email}
\usage{
send_email_tls(study_type, dest, attachment_list)
}
\arguments{
\item{study_type}{String containing the type of study to specific the email subject}

\item{dest}{Vector of strings containing email addresses}

\item{attachment_list}{Vector containing the paths to the files to be attached to the email}
}
\description{
This function sends an automated email with the generated reports.
The gmail() function is only available for emayili v0.6.5 and above.
It uses by default port 587 (TLS), but the port argument can be specified (e.g., port 465 for SSL).
}
