% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_weekly_email}
\alias{send_weekly_email}
\title{Send email}
\usage{
send_weekly_email(study_type, dest, attachment_list)
}
\arguments{
\item{study_type}{String containing the type of study to specific the email subject}

\item{dest}{Vector of strings containing email addresses}

\item{attachment_list}{Vector containing the paths to the files to be attached to the email}
}
\description{
This function sends an automated email with the generated reports
}
