% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{set_language_settings}
\alias{set_language_settings}
\title{Set language settings}
\usage{
set_language_settings()
}
\description{
This function sets country-specific language preferences to be used in the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
The function more specifically sets the environment variable `LANG` and teh current locale `LC_TIME`
}
