% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{set_odkc_connection}
\alias{set_odkc_connection}
\title{Set the connection to ODK Central through ruODK}
\usage{
set_odkc_connection(verbose = FALSE)
}
\arguments{
\item{verbose}{Boolean (optional) that enables (TRUE) or disables (FALSE) verbose output. Defaut is set to FALSE.}
}
\value{
list of ODK central project IDs
}
\description{
This function sets the connection to the country ODK Central server which stores data in the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
To this aim, the function relies on the `ruODK` package which is an R client to access and parse data from ODK Central.
}
