% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_from_xls_dict}
\alias{subset_from_xls_dict}
\title{Subset columns using an external Excel dictionary}
\usage{
subset_from_xls_dict(df, xls_dict)
}
\arguments{
\item{df}{Input dataframe}

\item{xls_dict}{Excel file containing 2 columns ('old' and 'new') that map the names of the variables in the input dataframe and the names of the variables in the output dataframe}
}
\value{
This function returns a dataframe.
}
\description{
Subset columns using an external Excel dictionary
}
