% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{summary_table_disp}
\alias{summary_table_disp}
\title{Process weekly facility assessment data (TIMCI-specific function)}
\usage{
summary_table_disp(all, varname, gname)
}
\arguments{
\item{all}{dataframe}

\item{varname}{variable name}

\item{gname}{group name}
}
\value{
This function returns a summary table with.
}
\description{
Process weekly facility assessment data (TIMCI-specific function)
}
