% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{text_2_plot}
\alias{text_2_plot}
\title{Plot string in image}
\usage{
text_2_plot(string)
}
\arguments{
\item{string}{String to be converted to an image}
}
\value{
This function returns a plot
}
\description{
Plot string in image
}
