% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_odkapi.R
\name{upload_odkc_csv_attachment}
\alias{upload_odkc_csv_attachment}
\title{Upload the new CSV created as a form attachment
See reference ODK API documentation in \url{https://odkcentral.docs.apiary.io/#reference/forms/draft-form/uploading-a-draft-form-attachment} for more details.}
\usage{
upload_odkc_csv_attachment(un, pw, form_url, csv_path)
}
\arguments{
\item{un}{string containing the ODK Central username.}

\item{pw}{string containing the password for user `un`.}

\item{form_url}{string containing the URL of the ODK Central form to be published, e.g. "https://sandbox.central.getodk.org/v1/projects/project-id/forms/form-id".}

\item{csv_path}{string containing the path to the CSV file to attach to the form.}
}
\value{
This function returns the response of the request to the ODK Central API.
}
\description{
Upload the new CSV created as a form attachment
See reference ODK API documentation in \url{https://odkcentral.docs.apiary.io/#reference/forms/draft-form/uploading-a-draft-form-attachment} for more details.
}
