% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xlsx_data_module.R
\name{xlsx_download_ui}
\alias{xlsx_download_ui}
\title{Excel data download module user interface}
\usage{
xlsx_download_ui(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\description{
Excel data download module user interface
}
