% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_name_consistency}
\alias{check_name_consistency}
\title{Check name consistency (TIMCI-specific function)}
\usage{
check_name_consistency(df, col_name1, col_name2)
}
\arguments{
\item{df}{dataframe}

\item{col_name1}{name of the column containing names in `df`}

\item{col_name2}{name of the column containing names in `df`}
}
\value{
This function returns a dataframe containing IDs and dates at which the ID has been allocated in different columns.
}
\description{
Check name consistency (TIMCI-specific function)
}
