% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{count_screening}
\alias{count_screening}
\title{Count the occurrence of a specific value in a column (TIMCI-specific function)}
\usage{
count_screening(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns de-identified data.
}
\description{
Count the occurrence of a specific value in a column (TIMCI-specific function)
}
