% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_day0_qc_flowchart}
\alias{create_day0_qc_flowchart}
\title{Create flowchart (TIMCI-specific)}
\usage{
create_day0_qc_flowchart(
  n_raw_records,
  n_nonvalid_deviceid_records,
  n_after_lockdate_records,
  n_screening,
  d,
  e,
  f,
  g
)
}
\arguments{
\item{n_raw_records}{Number of records}

\item{n_nonvalid_deviceid_records}{Number of records with a non-valid device ID}

\item{n_after_lockdate_records}{Number of Day 0 records entered after the lock date}

\item{n_screening}{TBD}

\item{d}{TBD}

\item{e}{TBD}

\item{f}{TBD}

\item{g}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create flowchart (TIMCI-specific)
}
