% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_day0_qc_flowchart}
\alias{create_day0_qc_flowchart}
\title{Create cleaning flowchart for Day 0 data (TIMCI-specific)}
\usage{
create_day0_qc_flowchart(
  n_raw_day0_records,
  n_nonvalid_fid_records,
  n_incorrect_enroldate_records,
  n_dropped_true_duplicate_records,
  n_duplicate_records,
  n_dropped_duplicate_records,
  n_edited_duplicate_records,
  n_drug_edits,
  n_negative_illness_onset,
  n_missing_cp,
  n_missing_diagnosis,
  n_missing_referral_cg,
  n_cleaned_day0_records
)
}
\arguments{
\item{n_raw_day0_records}{Initial number of Day 0 records}

\item{n_nonvalid_fid_records}{Number of Day 0 records with a non-valid device ID}

\item{n_incorrect_enroldate_records}{Number of Day 0 records with an incorrect enrolment date}

\item{n_dropped_duplicate_records}{Number of Day 0 records that were dropped}

\item{n_edited_duplicate_records}{Number of Day 0 records with a duplicated child ID that were edited}

\item{n_cleaned_day0_records}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for Day 0 data (TIMCI-specific)
}
