% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_day0_qc_flowchart}
\alias{create_day0_qc_flowchart}
\title{Create cleaning flowchart for Day 0 data (TIMCI-specific)}
\usage{
create_day0_qc_flowchart(
  n_raw_day0_records,
  n_nonvalid_fid_records,
  n_ineligible_cg_records,
  n_cleaned_day0_records,
  d,
  e,
  f,
  g
)
}
\arguments{
\item{n_raw_day0_records}{Number of records}

\item{n_nonvalid_fid_records}{Number of records with a non-valid device ID}

\item{n_ineligible_cg_records}{Number of Day 0 records with an entry date posterior to the lock date}

\item{n_cleaned_day0_records}{TBD}

\item{d}{TBD}

\item{e}{TBD}

\item{f}{TBD}

\item{g}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for Day 0 data (TIMCI-specific)
}
