% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_flowcharts.R
\name{create_day7fu_qc_flowchart}
\alias{create_day7fu_qc_flowchart}
\title{Create cleaning flowchart for Day 7 follow-up data (TIMCI-specific)}
\usage{
create_day7fu_qc_flowchart(
  n_raw_allday7fu_records,
  n_after_lockdate_records,
  n_nonvalid_pid_records,
  n_edit_nonvalid_pid_records,
  n_drop_nonvalid_pid_records,
  n_inconsistent_name_records,
  n_edit_inconsistent_name_records,
  n_drop_inconsistent_names_day7fu_records,
  n_cleaned_allday7fu_records
)
}
\arguments{
\item{n_raw_allday7fu_records}{Number of records}

\item{n_after_lockdate_records}{TBD}

\item{n_nonvalid_pid_records}{Number of records with a non-valid participan ID}

\item{n_cleaned_allday7fu_records}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create cleaning flowchart for Day 7 follow-up data (TIMCI-specific)
}
