% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_flowchart}
\alias{create_flowchart}
\title{Create flowchart (TIMCI-specific)}
\usage{
create_flowchart(n_records, n_screening, c, d, e, f, g)
}
\arguments{
\item{n_records}{Number of records}

\item{n_screening}{Number of screenings}

\item{c}{TBD}

\item{d}{TBD}

\item{e}{TBD}

\item{f}{TBD}

\item{g}{TBD}
}
\value{
This function returns a graph object
}
\description{
Create flowchart (TIMCI-specific)
}
