% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_modules.R
\name{csv_download_server}
\alias{csv_download_server}
\title{CSV data download module server-side processing}
\usage{
csv_download_server(id, name, df)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{name}{string containing the name of the data export}

\item{df}{parameter containing a dataframe}
}
\description{
CSV data download module server-side processing
}
