% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify_data}
\alias{deidentify_data}
\title{Deidentify data (TIMCI-specific function)}
\usage{
deidentify_data(df, mapping = NULL)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{mapping}{dataframe containing the mapping of the original / hashed IDs}
}
\value{
This function returns de-identified data.
}
\description{
De-identification of the TIMCI research data
}
