% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify_data.R
\name{deidentify_data}
\alias{deidentify_data}
\title{Deidentify data}
\usage{
deidentify_data(data)
}
\arguments{
\item{data}{numeric, the non de-identified (raw) data vector}
}
\value{
This function returns de-identified data.
}
\description{
De-identification of the TIMCI research data
}
