% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{deidentify_data}
\alias{deidentify_data}
\title{Deidentify data (TIMCI-specific function)}
\usage{
deidentify_data(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns de-identified data.
}
\description{
De-identification of the TIMCI research data
}
