% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timci_quality.R
\name{detect_inconsistent_dates}
\alias{detect_inconsistent_dates}
\title{Detect inconsistent dates (TIMCI-specific function)}
\usage{
detect_inconsistent_dates(df, col_date_start, col_date_end)
}
\arguments{
\item{df}{dataframe containing the processed facility data}

\item{col_date_start}{column}

\item{col_date_end}{column}
}
\value{
This function returns a dataframe containing data of possible duplicate participants only
}
\description{
Detect inconsistent dates (TIMCI-specific function)
}
