% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{display_weekly_fa_data_per_facility}
\alias{display_weekly_fa_data_per_facility}
\title{Display weekly facility assessment data per facility (TIMCI-specific function)}
\usage{
display_weekly_fa_data_per_facility(df, all_df)
}
\arguments{
\item{df}{dataframe containing the last available assessment for each facility}

\item{all_df}{dataframe containing all the processed facility assessment data}
}
\value{
This function outputs information about the weekly facility assessment data per facility.
}
\description{
Display weekly facility assessment data per facility (TIMCI-specific function)
}
