% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_timci_data.R
\name{edit_day0_child_ids}
\alias{edit_day0_child_ids}
\title{Edit incorrect child IDs in Day 0 data entries (TIMCI-specific function)
This function can be used to correct documented child ID duplicates, incorrect facility codes or typos}
\usage{
edit_day0_child_ids(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
This function returns a list that contains an edited dataframe and the list of edits
}
\description{
Edit incorrect child IDs in Day 0 data entries (TIMCI-specific function)
This function can be used to correct documented child ID duplicates, incorrect facility codes or typos
}
