% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_df2xlsx}
\alias{export_df2xlsx}
\title{Write dataframe to an Excel file, with a filename consisting of a prefix and a date stamp}
\usage{
export_df2xlsx(df, dirname, prefix, rnames = FALSE)
}
\arguments{
\item{df}{dataframe.}

\item{dirname}{directory where the Excel file will be created.}

\item{prefix}{filename prefix}

\item{rnames}{Row names}
}
\value{
This function returns the creation timestamp of the Excel file
}
\description{
Write dataframe to an Excel file, with a filename consisting of a prefix and a date stamp
}
