% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_screening_facility.R
\name{extract_daily_facility_for_deviceid}
\alias{extract_daily_facility_for_deviceid}
\title{Extract device IDs}
\usage{
extract_daily_facility_for_deviceid(facility_data, research_facilities)
}
\arguments{
\item{facility_data}{dataframe containing the processed facility data}

\item{research_facilities}{dataframe containing the list of research facilities and their characteristics}
}
\value{

}
\description{
Extract device IDs
}
