% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{extract_enrolled_participants}
\alias{extract_enrolled_participants}
\title{Extract enrolled participants}
\usage{
extract_enrolled_participants(df)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data}
}
\value{
This function returns a dataframe containing data of enrolled participants only
}
\description{
Extract enrolled participants
}
