% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{extract_hypoxaemia}
\alias{extract_hypoxaemia}
\title{Extract hypoxaemia}
\usage{
extract_hypoxaemia(df)
}
\arguments{
\item{df}{Dataframe containing the non deidentified facility data}
}
\value{
This function returns a dataframe containing data of hypoxemic children only
}
\description{
Extract hypoxaemia
}
