% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{extract_last_fa_data}
\alias{extract_last_fa_data}
\title{Extract the last available facility assessment data (TIMCI-specific function)}
\usage{
extract_last_fa_data(df)
}
\arguments{
\item{df}{dataframe containing the processed facility assessment data}
}
\value{
This function returns a dataframe that only contains the last available assessment for each facility.
}
\description{
Extract the last available facility assessment data (TIMCI-specific function)
}
