% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{extract_pii}
\alias{extract_pii}
\title{Extract Personally Identifiable Information (PII) (TIMCI-specific function)}
\usage{
extract_pii(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns de-identified data.
}
\description{
Extraction of the personally identifiable data from the data collected
}
