% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_data.R
\name{extract_referrals}
\alias{extract_referrals}
\title{Extract referrals (TIMCI-specific function)}
\usage{
extract_referrals(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing data of children who were referred at Day 0 only
}
\description{
Extract referrals (TIMCI-specific function)
}
