% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{extract_screening_data}
\alias{extract_screening_data}
\title{Extract screening data (TIMCI-specific function)}
\usage{
extract_screening_data(df)
}
\arguments{
\item{df}{dataframe containing the processed facility data}
}
\value{
This function returns a dataframe containing screening data only
}
\description{
Extract screening data (TIMCI-specific function)
}
