% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_odk_data.R
\name{format_odk_data}
\alias{format_odk_data}
\title{Format ODK data}
\usage{
format_odk_data(data)
}
\arguments{
\item{data}{ODK data}
}
\value{
This function formats ODK data.
}
\description{
Description of format_odk_data
}
