% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_calendar_heatmap}
\alias{generate_calendar_heatmap}
\title{Generate a calendar heatmap}
\usage{
generate_calendar_heatmap(df, datecol, date_max = Sys.Date(), date_min = NULL)
}
\arguments{
\item{df}{Data frame to use for the plot}

\item{datecol}{Column name in data frame `df` that contains dates}

\item{date_max}{End date of the plot (optional)}

\item{date_min}{Start date of the plot (optional)}
}
\value{
This function returns a ggplot object which contains a calendar heatmap.
}
\description{
Generate a calendar heatmap
}
