% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timci_data.R
\name{generate_cg_log}
\alias{generate_cg_log}
\title{Generate caregiver recruitment log (TIMCI-specific function)}
\usage{
generate_cg_log(pii, day7fu)
}
\arguments{
\item{pii}{dataframe containing personally identifiable data}

\item{day7fu}{dataframe containing day7 follow-up data}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of caregiver to be called for the qualitative studies
}
