% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_day_bar_plot}
\alias{generate_day_bar_plot}
\title{Create a daily bar plot}
\usage{
generate_day_bar_plot(date_vec, date_min, date_max)
}
\arguments{
\item{date_vec}{Vector containing dates}

\item{date_min}{Start date of the plot}

\item{date_max}{End date of the plot}
}
\value{
This function returns a ggplot object which contains a bar plot of frequencies by dates
}
\description{
generate_day_bar_plot() creates a bar plot.
}
