% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_enrolment_hist}
\alias{generate_enrolment_hist}
\title{Create a histogram that represents the distribution of values in `col2` for each category in `col1`}
\usage{
generate_enrolment_hist(df, col1, col2, lthres, uthres, lblx, lbly)
}
\arguments{
\item{df}{Data frame to use for the plot}

\item{col1}{Column name in data frame `df`}

\item{col2}{Column name in data frame `df` that contains numeric values}

\item{lthres}{Numeric value,  lower threshold}

\item{uthres}{Numeric value, upper threshold}

\item{lblx}{Label of the x-axis}

\item{lbly}{Label of the y-axis}
}
\value{
This function returns a ggplot object which contains a histogram representing the distribution of values in `col2` for each category in `col1`.
}
\description{
Red bars indicate health facilities which are far below the lower target.
Orange bars indicate health facilities which are above the lower target, but below the upper target.
Green bars indicate health facilities which are above the upper target.
}
