% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrolment_hist.R
\name{generate_enrolment_hist}
\alias{generate_enrolment_hist}
\title{Generate enrolment hist}
\usage{
generate_enrolment_hist(data)
}
\arguments{
\item{data}{Data}
}
\value{
This function generates a histogram representing enrolment in each facility.
}
\description{
This function plots a histogram representing enrolment.
}
\details{
Description of generate_enrolment_hist
}
