% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_enrolment_hist}
\alias{generate_enrolment_hist}
\title{Generate enrolment hist}
\usage{
generate_enrolment_hist(data)
}
\arguments{
\item{data}{Data}
}
\value{
This function returns a ggplot object of a histogram representing enrolment in each facility.
}
\description{
This function plots a histogram representing enrolment.
}
