% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{generate_fu_log}
\alias{generate_fu_log}
\title{Generate follow-up log}
\usage{
generate_fu_log(df, wmin, wmax)
}
\arguments{
\item{df}{dataframe}

\item{wmin}{numerical, start of the follow-up period (in days)}

\item{wmax}{numerical, end of the follow-up period (in days)}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants to be called in a time window after baseline between wmin and wmax
}
