% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fu_data.R
\name{generate_fu_log}
\alias{generate_fu_log}
\title{Generate follow-up log (TIMCI-specific function)}
\usage{
generate_fu_log(
  pii,
  fudf,
  wmin,
  wmax,
  vwmin,
  vwmax,
  ext = FALSE,
  deidentify = FALSE
)
}
\arguments{
\item{pii}{dataframe containing personally identifiable data}

\item{fudf}{dataframe containing the processed follow-up data}

\item{wmin}{numerical, start of the follow-up period (in days)}

\item{wmax}{numerical, end of the follow-up period (in days)}

\item{vwmin}{numerical, start of the follow-up period valid for the statistical analysis (in days)}

\item{vwmax}{numerical, end of the follow-up period valid for the statistical analysis (in days)}

\item{ext}{boolean, allow to export more information}

\item{deidentify}{boolean, enable the generation of a format with personally identifiable information (PII) for upload on ODK Central (if set to FALSE) or of a deidentified dataframe (if set to TRUE)}
}
\value{
This function returns a dataframe.
}
\description{
Generate a list of participants to be called in a time window after baseline between wmin and wmax
}
