% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkwdown_utils.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Generate a PDF report using Rmarkdown}
\usage{
generate_pdf_report(report_dir, rmd_fn, report_fn, rmd_params = list())
}
\arguments{
\item{report_dir}{Path to the folder where the generated PDF Rmarkdown report is stored}

\item{rmd_fn}{Filename of the Rmarkdown file}

\item{report_fn}{Filename of the Rmarkdown rendered report}

\item{rmd_params}{List of parameters}
}
\description{
This function generates a standardised Rmarkdown report in a PDF format
}
