% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_pie_chart}
\alias{generate_pie_chart}
\title{Create a pie chart plot}
\usage{
generate_pie_chart(df)
}
\arguments{
\item{df}{Dataframe to use for the plot}
}
\value{
This function returns a ggplot object which contains a pie chart.
}
\description{
generate_pie_chart() creates a pie chart plot.
}
