% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{generate_report}
\alias{generate_report}
\title{Generate Rmarkdown file}
\usage{
generate_report(report_dir, rmd_fn, report_fn, rmd_params = "")
}
\arguments{
\item{report_dir}{Path to the folder where the generated HTML and Microsoft Word DOCX Rmarkdown reports are stored}

\item{rmd_fn}{Filename of the Rmarkdown file}

\item{report_fn}{Filename of the Rmarkdown rendered report}

\item{rmd_params}{List of parameters}
}
\description{
This function generates a standardised Rmarkdown report in two formats HTML and Microsoft Word DOCX
}
