% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data_info_module.R
\name{odk_data_info_server}
\alias{odk_data_info_server}
\title{ODK data info module server-side processing}
\usage{
odk_data_info_server(id, df)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{df}{dataframe containing the current ODK data}
}
\description{
This module displays some global information about the ODK data table that is displayed.
}
