% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_data_table_server.R
\name{odk_data_table_server}
\alias{odk_data_table_server}
\title{Module UI function}
\usage{
odk_data_table_server(id, odk_form_list, odk_data)
}
\arguments{
\item{id}{Namespace for the Shiny module}

\item{odk_form_list}{ODK list of forms}

\item{odk_data}{ODK dataframe}
}
\description{
Module UI function
}
