% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enrolment_gauge}
\alias{plot_enrolment_gauge}
\title{Plot enrolment gauge}
\usage{
plot_enrolment_gauge(m, s, w, d)
}
\arguments{
\item{m}{Maximal gauge value}

\item{s}{Sucess}

\item{w}{Warning}

\item{d}{Danger}
}
\value{
This function plots a gauge representing enrolment.
}
\description{
This function plots a gauge representing enrolment.
}
\details{
Description of plot_enrolment_gauge
}
