% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_enrolment_gauge}
\alias{plot_enrolment_gauge}
\title{Create an enrolment gauge}
\usage{
plot_enrolment_gauge(df, m, s, w, d)
}
\arguments{
\item{df}{Dataframe}

\item{m}{Maximal gauge value}

\item{s}{Success}

\item{w}{Warning}

\item{d}{Danger}
}
\value{
This function plots a gauge representing enrolment.
}
\description{
plot_enrolment_gauge() creates a gauge plot that represents the global enrolment with respect to the target.
}
\details{
Description of plot_enrolment_gauge
}
