% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_odk_data.R
\name{process_facility_data}
\alias{process_facility_data}
\title{Process facility data}
\usage{
process_facility_data(df)
}
\arguments{
\item{df}{dataframe containing the non de-identified (raw) ODK data collected at the facility level}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process facility data
}
