% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_timci_fa_data.R
\name{process_weekly_fa_data}
\alias{process_weekly_fa_data}
\title{Process weekly facility assessment data (TIMCI-specific function)}
\usage{
process_weekly_fa_data(df, is_pilot = FALSE)
}
\arguments{
\item{df}{dataframe containing the (raw) ODK data collected in the weekly facility assessment}

\item{is_pilot}{Boolean that enables to select the pilot mode for Tanzania and India (optional, default set to FALSE)}
}
\value{
This function returns a formatted dataframe for future display and analysis.
}
\description{
Process weekly facility assessment data (TIMCI-specific function)
}
