% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown}
\alias{run_rmarkdown}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown(report_dir, participant_zip, db_dir, fu_dir, qual_dir)
}
\arguments{
\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{participant_zip}{Path to the encrypted zip archive that stores participant data}

\item{db_dir}{Path to the output folder for the database exports}

\item{fu_dir}{Path to the output folder for the follow-up exports}

\item{qual_dir}{Path to the output folder for the qualitative exports}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
