% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown}
\alias{run_rmarkdown}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown(
  report_dir,
  participant_zip,
  mdb_dir,
  fu_dir,
  qual1_dir,
  qual2_dir,
  spa_db_dir
)
}
\arguments{
\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{participant_zip}{Path to the encrypted zip archive that stores participant data}

\item{mdb_dir}{Path to the output folder for the RCT / LS database exports}

\item{fu_dir}{Path to the output folder for the follow-up exports}

\item{qual1_dir}{Path to the output folder for the caregiver IDI exports}

\item{qual2_dir}{Path to the output folder for the healthcare provider IDI exports}

\item{spa_db_dir}{Path to the output folder for the SPA database exports}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
