% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown_reportonly}
\alias{run_rmarkdown_reportonly}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown_reportonly(
  rctls_pid,
  rctls_pp,
  spa_pid,
  qpid,
  qual_pp,
  research_facilities,
  report_dir,
  participant_zip,
  mdb_dir,
  fu_dir,
  qc_dir,
  lock_dir,
  qualcg_dir,
  qualhcp_dir,
  spa_db_dir,
  path_dir,
  start_date = NULL,
  end_date = NULL,
  spa_pid2 = NULL,
  spa_start_date = NULL,
  lock_date = NULL,
  cost_pid = NULL,
  cost_dir = NULL,
  qualkii_dir = NULL,
  qualos_dir = NULL,
  sample_size = 1e+05,
  is_test = FALSE,
  operational_reports = TRUE,
  is_pilot = FALSE
)
}
\arguments{
\item{rctls_pid}{Numeric value that refers ID of the RCT/LS ODK Central project}

\item{rctls_pp}{Encryption passphrase associated with the RCT/LS ODK ODK Central project}

\item{spa_pid}{Numeric value that refers to the ID of the SPA / time-flow / process mapping project on the ODK Central server}

\item{qpid}{Numeric ID of the qualitative ODK Central project}

\item{qual_pp}{Encryption passphrase associated with the qualitative ODK Central project}

\item{research_facilities}{Dataframe that contains the research facilities}

\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{participant_zip}{Path to the encrypted zip archive that stores participant data}

\item{mdb_dir}{Path to the output folder for the RCT / LS database exports}

\item{fu_dir}{Path to the output folder for the follow-up exports}

\item{qc_dir}{Path to the output folder that contains the quality check findings}

\item{lock_dir}{Path to the output folder that contains the cleaned datasets}

\item{qualcg_dir}{Path to the output folder for the caregiver in-depth interview exports}

\item{qualhcp_dir}{Path to the output folder for the healthcare provider in-depth interview exports}

\item{spa_db_dir}{Path to the output folder for the SPA / time-flow / process mapping database exports}

\item{path_dir}{Path to the output folder for the M&E exports to be shared with PATH}

\item{start_date}{RCT/LS data collection start date (optional)}

\item{end_date}{RCT/LS data collection end date (optional)}

\item{spa_pid2}{Numeric value (optional) that refers to another ID of the SPA / time-flow / process mapping project on the ODK Central server if data have been split between 2 projects}

\item{spa_start_date}{SPA data collection start date (optional)}

\item{lock_date}{RCT/LS data collection cleaning end date for database lock (optional)}

\item{cost_pid}{Numeric value (optional) that refers to the ID of the cost and cost-effectiveness project on the ODK Central server}

\item{cost_dir}{Path to the output folder for the cost and cost effectiveness database exports (optional)}

\item{qualkii_dir}{Path to the output folder for the key informant interview exports (optional)}

\item{qualos_dir}{Path to the output folder for the online survey exports (optional)}

\item{sample_size}{Numeric value, sample size for RCT/LS enrolment}

\item{is_test}{Boolean that enables to export and process a small subset of data for running technical tests (optional, default set to FALSE)}

\item{operational_reports}{Boolean that enables to generate operational reports (optional, default set to TRUE)}

\item{is_pilot}{Boolean that enables to select the pilot mode for Tanzania and India (optional, default set to FALSE)}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
