% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rmarkdown.R
\name{run_rmarkdown_reportonly}
\alias{run_rmarkdown_reportonly}
\title{Run Rmarkdown files}
\usage{
run_rmarkdown_reportonly(
  rctls_pid,
  rctls_pp,
  spa_pid,
  qpid,
  qual_pp,
  research_facilities,
  report_dir,
  participant_zip,
  mdb_dir,
  fu_dir,
  qual1_dir,
  qual2_dir,
  spa_db_dir,
  path_dir,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{rctls_pid}{Numeric ID of the RCT/LS ODK Central project}

\item{rctls_pp}{Passphrase}

\item{spa_pid}{Numeric ID of the SPA / time-flow ODK Central project}

\item{qpid}{Numeric ID of the qualitative ODK Central project}

\item{qual_pp}{Passphrase}

\item{research_facilities}{Dataframe that contains the research facilities}

\item{report_dir}{Path to the output folder for the generated Rmarkdown reports}

\item{participant_zip}{Path to the encrypted zip archive that stores participant data}

\item{mdb_dir}{Path to the output folder for the RCT / LS database exports}

\item{fu_dir}{Path to the output folder for the follow-up exports}

\item{qual1_dir}{Path to the output folder for the caregiver IDI exports}

\item{qual2_dir}{Path to the output folder for the healthcare provider IDI exports}

\item{spa_db_dir}{Path to the output folder for the SPA database exports}

\item{path_dir}{Path to the output folder for the M&E exports to be shared with PATH}

\item{start_date}{start date}

\item{end_date}{end date}
}
\description{
This function runs several Rmarkdown files to generate standardised automated reports for the Tools for Integrated Management of Childhood Illnesses (TIMCI) project.
}
