% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_list_item.R
\name{select_list_item_server}
\alias{select_list_item_server}
\title{List item selection module server-side processing}
\usage{
select_list_item_server(input, output, session, list, data)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{data}{hash structure (non-reactive) containing ODK dataframes from different forms}

\item{clist}{list of items}
}
\value{
list with following components
\describe{
  \item{item}{reactive character string indicating item selection}
}
}
\description{
List item selection module server-side processing
}
