% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_list_item_module.R
\name{select_list_item_server}
\alias{select_list_item_server}
\title{List item selection module server-side processing}
\usage{
select_list_item_server(id, list, data)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{list}{list of items (non-reactive) containing ODK forms}

\item{data}{hash structure (non-reactive) containing the dataframes from different ODK forms}
}
\value{
reactive list with following components
\describe{
  \item{item}{reactive character string indicating item selection}
  \item{`cdata`}{dataframe containing the current ODK data}
}
}
\description{
List item selection module server-side processing
}
