% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xlsx_data_module.R
\name{xlsx_download_server}
\alias{xlsx_download_server}
\title{Excel data download module server-side processing}
\usage{
xlsx_download_server(id, vars)
}
\arguments{
\item{id}{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{vars}{parameter containing a reactive list with the following components:
\describe{
  \item{`citem`}{string indicating the current ODK form}
  \item{`cdata`}{dataframe containing the current ODK data}
}}
}
\description{
Excel data download module server-side processing
}
